refactor(server): split monolithic docker.py into modular package
Replace 2698-line docker.py with focused mixin modules:
- docker_service.py: core orchestration, CRUD, public API
- container_ops.py: image management, platform resolution, container creation
- networking.py: network validation, endpoints, egress sidecar
- volumes.py: volume validation, bind mounts, PVC
- runtime.py: execd distribution, bootstrap launcher
