--- server/opensandbox_server/services/docker.py
@@ -1,2698 +1,0 @@
-# Copyright 2025 Alibaba Group Holding Ltd.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""
-Docker-based implementation of SandboxService.
-
-This module provides a Docker implementation of the sandbox service interface,
-using Docker containers for sandbox lifecycle management.
-"""
-
-from __future__ import annotations
-
-import inspect
-import io
-import json
-import logging
-import math
-import os
-import posixpath
-import socket
-import tarfile
-import time
-from contextlib import contextmanager
-from dataclasses import dataclass
-from datetime import datetime, timedelta, timezone
-from threading import Lock, Timer
-from typing import Any, Dict, Optional
-from uuid import uuid4
-
-import docker
-from docker.errors import DockerException, ImageNotFound, NotFound as DockerNotFound
-from docker.types import DeviceRequest
-from fastapi import HTTPException, status
-
-from opensandbox_server.extensions import (
-    ACCESS_RENEW_EXTEND_SECONDS_METADATA_KEY,
-    apply_access_renew_extend_seconds_to_mapping,
-)
-from opensandbox_server.api.schema import (
-    CreateSandboxRequest,
-    CreateSandboxResponse,
-    Endpoint,
-    ImageSpec,
-    ListSandboxesRequest,
-    ListSandboxesResponse,
-    NetworkPolicy,
-    PaginationInfo,
-    RenewSandboxExpirationRequest,
-    RenewSandboxExpirationResponse,
-    Sandbox,
-    SandboxStatus,
-    PlatformSpec,
-)
-from opensandbox_server.config import AppConfig, get_config
-from opensandbox_server.services.docker_diagnostics import DockerDiagnosticsMixin
-from opensandbox_server.services.docker_port_allocator import (
-    allocate_port_bindings,
-    normalize_container_port_spec,
-    normalize_port_bindings,
-)
-from opensandbox_server.services.docker_windows_profile import (
-    apply_windows_runtime_host_config_defaults,
-    fetch_execd_install_bat,
-    fetch_execd_windows_binary,
-    inject_windows_resource_limits_env,
-    inject_windows_user_ports,
-    install_windows_oem_scripts,
-    is_windows_platform,
-    normalize_bootstrap_command,
-    resolve_docker_platform,
-    validate_windows_resource_limits,
-    validate_windows_runtime_prerequisites, WINDOWS_OEM_VOLUME_PREFIX,
-)
-from opensandbox_server.services.extension_service import ExtensionService
-from opensandbox_server.services.constants import (
-    EGRESS_MODE_ENV,
-    EGRESS_RULES_ENV,
-    OPENSANDBOX_EGRESS_TOKEN,
-    SANDBOX_EGRESS_AUTH_TOKEN_METADATA_KEY,
-    SANDBOX_EMBEDDING_PROXY_PORT_LABEL,
-    SANDBOX_EXPIRES_AT_LABEL,
-    SANDBOX_HTTP_PORT_LABEL,
-    SANDBOX_ID_LABEL,
-    SANDBOX_MANAGED_VOLUMES_LABEL,
-    SANDBOX_MANUAL_CLEANUP_LABEL,
-    SANDBOX_OSSFS_MOUNTS_LABEL,
-    SANDBOX_PLATFORM_ARCH_LABEL,
-    SANDBOX_PLATFORM_OS_LABEL,
-    SANDBOX_SNAPSHOT_ID_LABEL,
-    SandboxErrorCodes,
-)
-from opensandbox_server.services.endpoint_auth import (
-    build_egress_auth_headers,
-    generate_egress_token,
-    merge_endpoint_headers,
-)
-from opensandbox_server.services.helpers import (
-    matches_filter,
-    parse_gpu_request,
-    parse_memory_limit,
-    parse_nano_cpus,
-    parse_timestamp,
-)
-from opensandbox_server.services.ossfs_mixin import OSSFSMixin
-from opensandbox_server.services.sandbox_service import SandboxService
-from opensandbox_server.services.runtime_resolver import SecureRuntimeResolver
-from opensandbox_server.services.snapshot_restore import resolve_sandbox_image_from_request
-from opensandbox_server.services.validators import (
-    calculate_expiration_or_raise,
-    ensure_egress_configured,
-    ensure_entrypoint,
-    ensure_future_expiration,
-    ensure_metadata_labels,
-    ensure_platform_valid,
-    ensure_timeout_within_limit,
-    ensure_valid_host_path,
-    ensure_volumes_valid,
-)
-logger = logging.getLogger(__name__)
-
-
-def _running_inside_docker_container() -> bool:
-    """Return True if the current process is running inside a Docker container."""
-    return os.path.exists("/.dockerenv")
-
-
-OPENSANDBOX_DIR = "/opt/opensandbox"
-# Use posixpath for container-internal paths so they always use forward slashes,
-# even when the server runs on Windows.
-EXECED_INSTALL_PATH = posixpath.join(OPENSANDBOX_DIR, "execd")
-BOOTSTRAP_PATH = posixpath.join(OPENSANDBOX_DIR, "bootstrap.sh")
-
-HOST_NETWORK_MODE = "host"
-BRIDGE_NETWORK_MODE = "bridge"
-PENDING_FAILURE_TTL_SECONDS = int(os.environ.get("PENDING_FAILURE_TTL", "3600"))
-EGRESS_SIDECAR_LABEL = "opensandbox.io/egress-sidecar-for"
-
-
-@dataclass
-class PendingSandbox:
-    request: CreateSandboxRequest
-    created_at: datetime
-    expires_at: Optional[datetime]
-    status: SandboxStatus
-
-
-class DockerSandboxService(DockerDiagnosticsMixin, OSSFSMixin, SandboxService, ExtensionService):
-    """
-    Docker-based implementation of SandboxService.
-
-    This class implements sandbox lifecycle operations using Docker containers.
-    """
-
-    def __init__(self, config: Optional[AppConfig] = None):
-        """
-        Initialize Docker sandbox service.
-
-        Initializes Docker service from environment variables.
-        The service will read configuration from:
-        - DOCKER_HOST: Docker daemon URL (e.g., 'unix://var/run/docker.sock' or 'tcp://127.0.0.1:2376')
-        - DOCKER_TLS_CERTDIR: Directory containing TLS certificates
-        - Other Docker environment variables as needed
-
-        Note: Connection is not verified at initialization time.
-        Connection errors will be raised when Docker operations are performed.
-        """
-        self.app_config = config or get_config()
-        runtime_config = self.app_config.runtime
-        if runtime_config.type != "docker":
-            raise ValueError("DockerSandboxService requires runtime.type = 'docker'.")
-
-        self.execd_image = runtime_config.execd_image
-        self.network_mode = (self.app_config.docker.network_mode or HOST_NETWORK_MODE).lower()
-        self._execd_archive_cache: Dict[str, bytes] = {}
-        self._windows_profile_cache: Dict[str, bytes] = {}
-        self._daemon_platform: Optional[PlatformSpec] = None
-        self._api_timeout = self._resolve_api_timeout()
-        try:
-            # Initialize Docker service from environment variables
-            client_kwargs = {}
-            try:
-                signature = inspect.signature(docker.from_env)
-                if "timeout" in signature.parameters:
-                    client_kwargs["timeout"] = self._api_timeout
-            except (ValueError, TypeError):
-                logger.debug(
-                    "Unable to introspect docker.from_env signature; using default parameters."
-                )
-            self.docker_client = docker.from_env(**client_kwargs)
-            if not client_kwargs:
-                try:
-                    self.docker_client.api.timeout = self._api_timeout
-                except AttributeError:
-                    logger.debug("Docker client API does not expose timeout attribute.")
-            logger.info("Docker service initialized from environment")
-        except Exception as e:  # noqa: BLE001
-            # Common failure mode on macOS/dev machines: Docker daemon not running or socket path wrong.
-            hint = ""
-            msg = str(e)
-            if isinstance(e, FileNotFoundError) or "No such file or directory" in msg:
-                docker_host = os.environ.get("DOCKER_HOST", "")
-                hint = (
-                    " Docker daemon seems unavailable (unix socket not found). "
-                    "Make sure Docker Desktop (or Colima/Rancher Desktop) is running. "
-                    "If you use Colima on macOS, you may need to set "
-                    "DOCKER_HOST=unix://${HOME}/.colima/default/docker.sock before starting the server. "
-                    f"(current DOCKER_HOST='{docker_host}')"
-                )
-            raise HTTPException(
-                status_code=status.HTTP_503_SERVICE_UNAVAILABLE,
-                detail={
-                    "code": SandboxErrorCodes.DOCKER_INITIALIZATION_ERROR,
-                    "message": f"Failed to initialize Docker service: {str(e)}.{hint}",
-                },
-            )
-        self._expiration_lock = Lock()
-        self._execd_archive_lock = Lock()
-        self._sandbox_expirations: Dict[str, datetime] = {}
-        self._expiration_timers: Dict[str, Timer] = {}
-        self._pending_sandboxes: Dict[str, PendingSandbox] = {}
-        self._pending_lock = Lock()
-        self._pending_cleanup_timers: Dict[str, Timer] = {}
-        self._ossfs_mount_lock = Lock()
-        self._ossfs_mount_ref_counts: Dict[str, int] = {}
-        self._restore_existing_sandboxes()
-
-        # Initialize secure runtime resolver
-        self.resolver = SecureRuntimeResolver(self.app_config)
-        self.docker_runtime = self.resolver.get_docker_runtime()
-
-    def _resolve_api_timeout(self) -> int:
-        """Docker API timeout in seconds: [docker].api_timeout if set, else default 180."""
-        cfg = self.app_config.docker.api_timeout
-        if cfg is not None and cfg >= 1:
-            return cfg
-        return 180
-
-    @contextmanager
-    def _docker_operation(self, action: str, sandbox_id: Optional[str] = None):
-        """Context manager to log duration for Docker API calls."""
-        op_id = sandbox_id or "shared"
-        start = time.perf_counter()
-        try:
-            yield
-        except Exception as exc:
-            elapsed_ms = (time.perf_counter() - start) * 1000
-            logger.warning(
-                "sandbox=%s | action=%s | duration=%.2f | error=%s",
-                op_id,
-                action,
-                elapsed_ms,
-                exc,
-            )
-            raise
-        else:
-            elapsed_ms = (time.perf_counter() - start) * 1000
-            logger.info(
-                "sandbox=%s | action=%s | duration=%.2f",
-                op_id,
-                action,
-                elapsed_ms,
-            )
-
-    def _get_container_by_sandbox_id(self, sandbox_id: str):
-        """Helper to fetch the Docker container associated with a sandbox ID."""
-        label_selector = f"{SANDBOX_ID_LABEL}={sandbox_id}"
-        try:
-            containers = self.docker_client.containers.list(
-                all=True, filters={"label": label_selector}
-            )
-        except DockerException as exc:
-            raise HTTPException(
-                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                detail={
-                    "code": SandboxErrorCodes.CONTAINER_QUERY_FAILED,
-                    "message": f"Failed to query sandbox containers: {str(exc)}",
-                },
-            ) from exc
-
-        if not containers:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail={
-                    "code": SandboxErrorCodes.SANDBOX_NOT_FOUND,
-                    "message": f"Sandbox {sandbox_id} not found.",
-                },
-            )
-
-        return containers[0]
-
-    def _schedule_expiration(
-        self,
-        sandbox_id: str,
-        expires_at: datetime,
-        *,
-        update_expiration: bool = True,
-        **expire_kwargs,
-    ) -> None:
-        """Schedule automatic sandbox termination at expiration time."""
-        # Delay might already be negative if the timer should fire immediately
-        delay = max(0.0, (expires_at - datetime.now(timezone.utc)).total_seconds())
-        timer = Timer(
-            delay,
-            self._expire_sandbox,
-            args=(sandbox_id,),
-            kwargs=expire_kwargs or None,
-        )
-        timer.daemon = True
-        with self._expiration_lock:
-            # Replace existing timer (if any) so renew operations take effect immediately
-            existing = self._expiration_timers.pop(sandbox_id, None)
-            if existing:
-                existing.cancel()
-            if update_expiration:
-                self._sandbox_expirations[sandbox_id] = expires_at
-            self._expiration_timers[sandbox_id] = timer
-        timer.start()
-
-    def _remove_expiration_tracking(self, sandbox_id: str) -> None:
-        """Remove expiration tracking and cancel any pending timers."""
-        with self._expiration_lock:
-            timer = self._expiration_timers.pop(sandbox_id, None)
-            if timer:
-                timer.cancel()
-            self._sandbox_expirations.pop(sandbox_id, None)
-
-    @staticmethod
-    def _has_manual_cleanup(labels: Dict[str, str]) -> bool:
-        """Return True when labels indicate manual cleanup mode."""
-        return labels.get(SANDBOX_MANUAL_CLEANUP_LABEL, "").lower() == "true"
-
-    def _get_tracked_expiration(
-        self,
-        sandbox_id: str,
-        labels: Dict[str, str],
-    ) -> Optional[datetime]:
-        """Return the known expiration timestamp for the sandbox."""
-        with self._expiration_lock:
-            tracked = self._sandbox_expirations.get(sandbox_id)
-        if tracked:
-            return tracked
-        label_value = labels.get(SANDBOX_EXPIRES_AT_LABEL)
-        if label_value:
-            return parse_timestamp(label_value)
-        return None
-
-    def _expire_sandbox(
-        self,
-        sandbox_id: str,
-        fallback_mount_keys: Optional[list[str]] = None,
-    ) -> None:
-        """Timer callback to terminate expired sandboxes."""
-        mount_keys: list[str] = []
-        try:
-            container = self._get_container_by_sandbox_id(sandbox_id)
-        except HTTPException as exc:
-            if exc.status_code == status.HTTP_404_NOT_FOUND:
-                self._remove_expiration_tracking(sandbox_id)
-                self._cleanup_windows_oem_volume(sandbox_id, None)
-                if fallback_mount_keys:
-                    self._release_ossfs_mounts(fallback_mount_keys)
-            else:
-                with self._expiration_lock:
-                    current_expires = self._sandbox_expirations.get(sandbox_id)
-                now = datetime.now(timezone.utc)
-                if current_expires and current_expires > now:
-                    logger.info(
-                        "Sandbox %s expiration was renewed; skipping retry.",
-                        sandbox_id,
-                    )
-                else:
-                    logger.warning(
-                        "Failed to fetch sandbox %s for expiration: %s — "
-                        "scheduling retry in 30s",
-                        sandbox_id,
-                        exc.detail,
-                    )
-                    retry_at = now + timedelta(seconds=30)
-                    self._schedule_expiration(
-                        sandbox_id,
-                        retry_at,
-                        update_expiration=False,
-                        fallback_mount_keys=fallback_mount_keys,
-                    )
-            return
-
-        with self._expiration_lock:
-            current_expires = self._sandbox_expirations.get(sandbox_id)
-        if current_expires and current_expires > datetime.now(timezone.utc):
-            logger.info(
-                "Sandbox %s was renewed (expires %s); aborting expiration.",
-                sandbox_id,
-                current_expires,
-            )
-            return
-
-        labels = container.attrs.get("Config", {}).get("Labels") or {}
-        mount_keys_raw = labels.get(SANDBOX_OSSFS_MOUNTS_LABEL, "[]")
-        try:
-            parsed_mount_keys = json.loads(mount_keys_raw)
-            if isinstance(parsed_mount_keys, list):
-                mount_keys = [key for key in parsed_mount_keys if isinstance(key, str) and key]
-        except (TypeError, json.JSONDecodeError):
-            mount_keys = []
-
-        try:
-            state = container.attrs.get("State", {})
-            if state.get("Running", False):
-                container.kill()
-        except DockerException as exc:
-            logger.warning("Failed to stop expired sandbox %s: %s", sandbox_id, exc)
-
-        try:
-            container.remove(force=True)
-        except DockerException as exc:
-            logger.warning("Failed to remove expired sandbox %s: %s", sandbox_id, exc)
-
-        self._remove_expiration_tracking(sandbox_id)
-        # Ensure sidecar is also cleaned up on expiration
-        self._cleanup_egress_sidecar(sandbox_id)
-        self._cleanup_windows_oem_volume(sandbox_id, labels)
-        self._release_ossfs_mounts(mount_keys)
-
-    def _restore_existing_sandboxes(self) -> None:
-        """On startup, rebuild expiration timers for containers already running."""
-        try:
-            containers = self.docker_client.containers.list(all=True)
-        except DockerException as exc:
-            logger.warning("Failed to restore existing sandboxes: %s", exc)
-            return
-
-        restored = 0
-        seen_sidecars: set[str] = set()
-        restored_mount_refs: dict[str, int] = {}
-        expired_entries: list[tuple[str, list[str]]] = []
-        now = datetime.now(timezone.utc)
-
-        def _parse_and_accumulate_mount_refs(labels: dict) -> list[str]:
-            mount_keys_raw = labels.get(SANDBOX_OSSFS_MOUNTS_LABEL, "[]")
-            try:
-                parsed = json.loads(mount_keys_raw)
-            except (TypeError, json.JSONDecodeError):
-                parsed = []
-            keys: list[str] = []
-            if isinstance(parsed, list):
-                for key in parsed:
-                    if isinstance(key, str) and key:
-                        keys.append(key)
-                        restored_mount_refs[key] = restored_mount_refs.get(key, 0) + 1
-            return keys
-
-        # Pass 1: collect ref counts for ALL sandbox containers (alive + expired)
-        # and schedule timers for alive ones.  Expired sandboxes are deferred to
-        # pass 2 so that ref counts are fully populated before any release.
-        for container in containers:
-            labels = container.attrs.get("Config", {}).get("Labels") or {}
-            sidecar_for = labels.get(EGRESS_SIDECAR_LABEL)
-            if sidecar_for:
-                seen_sidecars.add(sidecar_for)
-                continue
-
-            sandbox_id = labels.get(SANDBOX_ID_LABEL)
-            if not sandbox_id:
-                continue
-
-            mount_keys = _parse_and_accumulate_mount_refs(labels)
-
-            expires_label = labels.get(SANDBOX_EXPIRES_AT_LABEL)
-            if expires_label:
-                expires_at = parse_timestamp(expires_label)
-            elif self._has_manual_cleanup(labels):
-                restored += 1
-                continue
-            else:
-                logger.warning(
-                    "Sandbox %s missing expires-at label; skipping expiration scheduling.",
-                    sandbox_id,
-                )
-                continue
-
-            if expires_at <= now:
-                logger.info("Sandbox %s already expired; terminating now.", sandbox_id)
-                expired_entries.append((sandbox_id, mount_keys))
-                continue
-
-            self._schedule_expiration(sandbox_id, expires_at)
-            restored += 1
-
-        # Populate ref counts before expiring anything so _release_ossfs_mount
-        # can properly decrement and unmount.
-        with self._ossfs_mount_lock:
-            self._ossfs_mount_ref_counts = restored_mount_refs
-
-        # Pass 2: expire deferred sandboxes (ref counts are now available).
-        # Cached mount keys are passed as fallback so that mounts are still
-        # released even if the container vanishes between pass 1 and pass 2.
-        for sandbox_id, cached_mount_keys in expired_entries:
-            self._expire_sandbox(sandbox_id, fallback_mount_keys=cached_mount_keys)
-
-        # Cleanup orphan sidecars (no matching sandbox container)
-        for orphan_id in seen_sidecars:
-            try:
-                self._get_container_by_sandbox_id(orphan_id)
-            except HTTPException as exc:
-                if exc.status_code == status.HTTP_404_NOT_FOUND:
-                    self._cleanup_egress_sidecar(orphan_id)
-                else:
-                    logger.warning(
-                        "Failed to check sandbox %s for orphan sidecar cleanup: %s", orphan_id, exc
-                    )
-
-        if restored:
-            logger.info("Restored expiration timers for %d sandbox(es).", restored)
-
-    def _normalize_platform_key(self, platform: Optional[PlatformSpec]) -> str:
-        if platform is None:
-            return "default"
-        return f"{platform.os}/{platform.arch}"
-
-    @staticmethod
-    def _normalize_arch(arch: Optional[str]) -> Optional[str]:
-        if not isinstance(arch, str):
-            return None
-        normalized = arch.strip().lower()
-        arch_aliases = {
-            "x86_64": "amd64",
-            "x86-64": "amd64",
-            "amd64": "amd64",
-            "aarch64": "arm64",
-            "arm64/v8": "arm64",
-            "arm64v8": "arm64",
-            "arm64": "arm64",
-        }
-        return arch_aliases.get(normalized, normalized)
-
-    @staticmethod
-    def _normalize_os(os_value: Optional[str]) -> Optional[str]:
-        if not isinstance(os_value, str):
-            return None
-        normalized = os_value.strip().lower()
-        os_aliases = {
-            "linux": "linux",
-        }
-        return os_aliases.get(normalized, normalized)
-
-    def _get_daemon_platform(self) -> Optional[PlatformSpec]:
-        if self._daemon_platform is not None:
-            return self._daemon_platform
-        try:
-            info = self.docker_client.info() or {}
-        except DockerException as exc:
-            logger.debug("Failed to inspect Docker daemon platform: %s", exc)
-            return None
-        os_value = info.get("OSType") or info.get("Os") or info.get("os")
-        arch_value = info.get("Architecture") or info.get("architecture")
-        if not isinstance(os_value, str) or not isinstance(arch_value, str):
-            return None
-        normalized_os = self._normalize_os(os_value)
-        normalized_arch = self._normalize_arch(arch_value)
-        if not normalized_os or not normalized_arch:
-            return None
-        self._daemon_platform = PlatformSpec(os=normalized_os, arch=normalized_arch)
-        return self._daemon_platform
-
-    def _fetch_execd_archive(self, platform: Optional[PlatformSpec] = None) -> bytes:
-        """Fetch (and memoize) the execd archive by effective target platform."""
-        cache_key = self._normalize_platform_key(platform)
-        if cache_key in self._execd_archive_cache:
-            return self._execd_archive_cache[cache_key]
-
-        with self._execd_archive_lock:
-            # Double-check locking to ensure only one thread initializes the cache
-            if cache_key in self._execd_archive_cache:
-                return self._execd_archive_cache[cache_key]
-
-            container = None
-            docker_platform = None
-            if platform is not None:
-                docker_platform = f"{platform.os}/{platform.arch}"
-            try:
-                self._ensure_image_available(
-                    self.execd_image,
-                    auth_config=None,
-                    sandbox_id=f"execd-cache:{cache_key}",
-                    platform=platform,
-                )
-
-                with self._docker_operation("execd cache create container", "execd-cache"):
-                    create_kwargs: dict[str, Any] = {
-                        "image": self.execd_image,
-                        "command": ["tail", "-f", "/dev/null"],
-                        "name": f"sandbox-execd-{uuid4()}",
-                        "detach": True,
-                        "auto_remove": False,
-                    }
-                    if docker_platform is not None:
-                        create_kwargs["platform"] = docker_platform
-                    container = self.docker_client.containers.create(**create_kwargs)
-                with self._docker_operation("execd cache start container", "execd-cache"):
-                    container.start()
-                    container.reload()
-                    logger.info("Created sandbox execd archive for container %s", container.id)
-            except TypeError as exc:
-                if docker_platform is not None:
-                    raise HTTPException(
-                        status_code=status.HTTP_400_BAD_REQUEST,
-                        detail={
-                            "code": SandboxErrorCodes.INVALID_PARAMETER,
-                            "message": (
-                                "The configured Docker client/daemon does not support "
-                                f"platform-aware container create for '{docker_platform}'."
-                            ),
-                        },
-                    ) from exc
-                raise
-            except DockerException as exc:
-                raise HTTPException(
-                    status_code=status.HTTP_503_SERVICE_UNAVAILABLE,
-                    detail={
-                        "code": SandboxErrorCodes.EXECD_START_FAILED,
-                        "message": f"Failed to start execd container: {str(exc)}",
-                    },
-                ) from exc
-
-            try:
-                with self._docker_operation("execd cache read archive", "execd-cache"):
-                    stream, _ = container.get_archive("/execd")
-                    data = b"".join(stream)
-            except DockerException as exc:
-                raise HTTPException(
-                    status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                    detail={
-                        "code": SandboxErrorCodes.EXECD_DISTRIBUTION_FAILED,
-                        "message": f"Failed to read execd artifacts: {str(exc)}",
-                    },
-                ) from exc
-            finally:
-                if container:
-                    try:
-                        with self._docker_operation("execd cache cleanup container", "execd-cache"):
-                            container.remove(force=True)
-                    except DockerException as cleanup_exc:
-                        logger.warning(
-                            "Failed to cleanup temporary execd container: %s", cleanup_exc
-                        )
-
-            self._execd_archive_cache[cache_key] = data
-            logger.info("Dumped execd archive to memory for platform key %s", cache_key)
-            return data
-
-    def _container_to_sandbox(self, container, sandbox_id: Optional[str] = None) -> Sandbox:
-        labels = container.attrs.get("Config", {}).get("Labels") or {}
-        resolved_id = sandbox_id or labels.get(SANDBOX_ID_LABEL)
-        if not resolved_id:
-            raise HTTPException(
-                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                detail={
-                    "code": SandboxErrorCodes.SANDBOX_NOT_FOUND,
-                    "message": "Container missing sandbox ID label.",
-                },
-            )
-
-        status_section = container.attrs.get("State", {})
-        status_value = (status_section.get("Status") or container.status or "").lower()
-        running = status_section.get("Running", False)
-        paused = status_section.get("Paused", False)
-        restarting = status_section.get("Restarting", False)
-        exit_code = status_section.get("ExitCode")
-        finished_at = status_section.get("FinishedAt")
-
-        if running and not paused:
-            state = "Running"
-            reason = "CONTAINER_RUNNING"
-            message = "Sandbox container is running."
-        elif paused:
-            state = "Paused"
-            reason = "CONTAINER_PAUSED"
-            message = "Sandbox container is paused."
-        elif restarting:
-            state = "Running"
-            reason = "CONTAINER_RESTARTING"
-            message = "Sandbox container is restarting."
-        elif status_value in {"created", "starting"}:
-            state = "Pending"
-            reason = "CONTAINER_STARTING"
-            message = "Sandbox container is starting."
-        elif status_value in {"exited", "dead"}:
-            if exit_code == 0:
-                state = "Terminated"
-                reason = "CONTAINER_EXITED"
-                message = "Sandbox container exited successfully."
-            else:
-                state = "Failed"
-                reason = "CONTAINER_EXITED_ERROR"
-                message = f"Sandbox container exited with code {exit_code}."
-        else:
-            state = "Unknown"
-            reason = "CONTAINER_STATE_UNKNOWN"
-            message = f"Sandbox container is in state '{status_value or 'unknown'}'."
-
-        metadata = {
-            key: value
-            for key, value in labels.items()
-            if key
-            not in {
-                SANDBOX_ID_LABEL,
-                SANDBOX_EXPIRES_AT_LABEL,
-                SANDBOX_MANUAL_CLEANUP_LABEL,
-                SANDBOX_PLATFORM_OS_LABEL,
-                SANDBOX_PLATFORM_ARCH_LABEL,
-                SANDBOX_SNAPSHOT_ID_LABEL,
-                ACCESS_RENEW_EXTEND_SECONDS_METADATA_KEY,
-            }
-        } or None
-        entrypoint = container.attrs.get("Config", {}).get("Cmd") or []
-        if isinstance(entrypoint, str):
-            entrypoint = [entrypoint]
-        image_tags = container.image.tags
-        image_uri = image_tags[0] if image_tags else container.image.short_id
-        snapshot_id = labels.get(SANDBOX_SNAPSHOT_ID_LABEL)
-        image_spec = None if snapshot_id else ImageSpec(uri=image_uri)
-
-        created_at = parse_timestamp(container.attrs.get("Created"))
-        last_transition_at = (
-            parse_timestamp(finished_at)
-            if finished_at and finished_at != "0001-01-01T00:00:00Z"
-            else created_at
-        )
-        expires_at = self._get_tracked_expiration(resolved_id, labels)
-
-        status_info = SandboxStatus(
-            state=state,
-            reason=reason,
-            message=message,
-            last_transition_at=last_transition_at,
-        )
-        platform_spec = self._resolve_platform_for_container(container, labels)
-
-        return Sandbox(
-            id=resolved_id,
-            image=image_spec,
-            snapshotId=snapshot_id,
-            platform=platform_spec,
-            status=status_info,
-            metadata=metadata,
-            entrypoint=entrypoint,
-            expiresAt=expires_at,
-            createdAt=created_at,
-        )
-
-    @staticmethod
-    def _platform_from_labels(labels: Dict[str, str]) -> Optional[PlatformSpec]:
-        os_value = labels.get(SANDBOX_PLATFORM_OS_LABEL)
-        arch_value = labels.get(SANDBOX_PLATFORM_ARCH_LABEL)
-        if not isinstance(os_value, str) or not isinstance(arch_value, str):
-            return None
-        if not os_value or not arch_value:
-            return None
-        normalized_os = DockerSandboxService._normalize_os(os_value)
-        normalized_arch = DockerSandboxService._normalize_arch(arch_value)
-        if not normalized_os or not normalized_arch:
-            return None
-        return PlatformSpec(os=normalized_os, arch=normalized_arch)
-
-    def _resolve_platform_for_container(
-        self,
-        container,
-        labels: Dict[str, str],
-        include_runtime_metadata: bool = False,
-    ) -> Optional[PlatformSpec]:
-        # API contract: platform is only echoed from explicit constraints.
-        # Runtime image metadata is used only for internal runtime preparation.
-        if include_runtime_metadata:
-            image_attrs = getattr(container.image, "attrs", {}) or {}
-            if isinstance(image_attrs, dict):
-                os_value = image_attrs.get("Os") or image_attrs.get("os")
-                arch_value = image_attrs.get("Architecture") or image_attrs.get("architecture")
-            else:
-                os_value = None
-                arch_value = None
-            if isinstance(os_value, str) and isinstance(arch_value, str) and os_value and arch_value:
-                normalized_os = self._normalize_os(os_value)
-                normalized_arch = self._normalize_arch(arch_value)
-                if normalized_os and normalized_arch:
-                    return PlatformSpec(os=normalized_os, arch=normalized_arch)
-        return self._platform_from_labels(labels)
-
-    def _ensure_directory(self, container, path: str, sandbox_id: Optional[str] = None) -> None:
-        """Create a directory within the target container if it does not exist."""
-        if not path or path == "/":
-            return
-        normalized_path = path.rstrip("/")
-        if not normalized_path:
-            return
-        tar_stream = io.BytesIO()
-        with tarfile.open(fileobj=tar_stream, mode="w") as tar:
-            dir_info = tarfile.TarInfo(name=normalized_path.lstrip("/"))
-            dir_info.type = tarfile.DIRTYPE
-            dir_info.mode = 0o755
-            dir_info.mtime = int(time.time())
-            tar.addfile(dir_info)
-        tar_stream.seek(0)
-        try:
-            with self._docker_operation(f"ensure directory {normalized_path}", sandbox_id):
-                container.put_archive(path="/", data=tar_stream.getvalue())
-        except DockerException as exc:
-            raise HTTPException(
-                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                detail={
-                    "code": SandboxErrorCodes.EXECD_DISTRIBUTION_FAILED,
-                    "message": f"Failed to create directory {path} in sandbox: {str(exc)}",
-                },
-            ) from exc
-
-    def _copy_execd_to_container(
-        self,
-        container,
-        sandbox_id: str,
-        platform: Optional[PlatformSpec] = None,
-    ) -> None:
-        """Copy execd artifacts from the platform container into the sandbox."""
-        archive = self._fetch_execd_archive(platform)
-        target_parent = posixpath.dirname(EXECED_INSTALL_PATH.rstrip("/")) or "/"
-        self._ensure_directory(container, target_parent, sandbox_id)
-        try:
-            with self._docker_operation("copy execd archive to sandbox", sandbox_id):
-                container.put_archive(path=target_parent, data=archive)
-        except DockerException as exc:
-            raise HTTPException(
-                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                detail={
-                    "code": SandboxErrorCodes.EXECD_DISTRIBUTION_FAILED,
-                    "message": f"Failed to copy execd into sandbox: {str(exc)}",
-                },
-            ) from exc
-
-    def _install_bootstrap_script(self, container, sandbox_id: str) -> None:
-        """Install the bootstrap launcher that starts execd then chains to user command."""
-        script_path = BOOTSTRAP_PATH
-        script_dir = posixpath.dirname(script_path)
-        self._ensure_directory(container, script_dir, sandbox_id)
-        execd_binary = EXECED_INSTALL_PATH
-        script_content = "\n".join(
-            [
-                "#!/bin/sh",
-                "set -e",
-                f"  {execd_binary} >/tmp/execd.log 2>&1 &",
-                'exec "$@"',
-                "",
-            ]
-        ).encode("utf-8")
-
-        tar_stream = io.BytesIO()
-        with tarfile.open(fileobj=tar_stream, mode="w") as tar:
-            info = tarfile.TarInfo(name=script_path.lstrip("/"))
-            info.mode = 0o755
-            info.size = len(script_content)
-            info.mtime = int(time.time())
-            tar.addfile(info, io.BytesIO(script_content))
-        tar_stream.seek(0)
-        try:
-            with self._docker_operation("install bootstrap script", sandbox_id):
-                container.put_archive(path="/", data=tar_stream.getvalue())
-        except DockerException as exc:
-            raise HTTPException(
-                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                detail={
-                    "code": SandboxErrorCodes.BOOTSTRAP_INSTALL_FAILED,
-                    "message": f"Failed to install bootstrap launcher: {str(exc)}",
-                },
-            ) from exc
-
-    def _prepare_sandbox_runtime(
-        self,
-        container,
-        sandbox_id: str,
-        platform: Optional[PlatformSpec] = None,
-    ) -> None:
-        """Copy execd artifacts and bootstrap launcher into the sandbox container."""
-        self._copy_execd_to_container(container, sandbox_id, platform)
-        self._install_bootstrap_script(container, sandbox_id)
-
-    def _prepare_creation_context(
-        self,
-        request: CreateSandboxRequest,
-    ) -> tuple[str, datetime, Optional[datetime]]:
-        sandbox_id = self.generate_sandbox_id()
-        created_at = datetime.now(timezone.utc)
-        expires_at = None
-        if request.timeout is not None:
-            expires_at = calculate_expiration_or_raise(created_at, request.timeout)
-        return sandbox_id, created_at, expires_at
-
-    async def create_sandbox(self, request: CreateSandboxRequest) -> CreateSandboxResponse:
-        """
-        Create a new sandbox from a container image using Docker.
-
-        Args:
-            request: Sandbox creation request
-
-        Returns:
-            CreateSandboxResponse: Created sandbox information
-
-        Raises:
-            HTTPException: If sandbox creation fails
-        """
-        request = resolve_sandbox_image_from_request(request)
-        ensure_entrypoint(request.entrypoint or [])
-        ensure_metadata_labels(request.metadata)
-        ensure_platform_valid(request.platform)
-        ensure_timeout_within_limit(
-            request.timeout,
-            self.app_config.server.max_sandbox_timeout_seconds,
-        )
-        self._ensure_secure_access_support(request)
-        self._ensure_network_policy_support(request)
-        self._validate_network_exists()
-        pvc_inspect_cache, auto_created_volumes = self._validate_volumes(request)
-        sandbox_id, created_at, expires_at = self._prepare_creation_context(request)
-        return self._provision_sandbox(
-            sandbox_id, request, created_at, expires_at, pvc_inspect_cache, auto_created_volumes,
-        )
-
-    def _async_provision_worker(
-        self,
-        sandbox_id: str,
-        request: CreateSandboxRequest,
-        created_at: datetime,
-        expires_at: Optional[datetime],
-        pvc_inspect_cache: Optional[dict[str, dict]] = None,
-    ) -> None:
-        try:
-            self._provision_sandbox(sandbox_id, request, created_at, expires_at, pvc_inspect_cache)
-        except HTTPException as exc:
-            message = exc.detail.get("message") if isinstance(exc.detail, dict) else str(exc)
-            self._mark_pending_failed(sandbox_id, message or "Sandbox provisioning failed.")
-            self._cleanup_failed_containers(sandbox_id)
-            self._schedule_pending_cleanup(sandbox_id)
-        except Exception as exc:  # noqa: BLE001
-            logger.exception("Unexpected error provisioning sandbox %s: %s", sandbox_id, exc)
-            self._mark_pending_failed(sandbox_id, str(exc))
-            self._cleanup_failed_containers(sandbox_id)
-            self._schedule_pending_cleanup(sandbox_id)
-        else:
-            self._remove_pending_sandbox(sandbox_id)
-
-    def _mark_pending_failed(self, sandbox_id: str, message: str) -> None:
-        with self._pending_lock:
-            pending = self._pending_sandboxes.get(sandbox_id)
-            if not pending:
-                return
-            pending.status = SandboxStatus(
-                state="Failed",
-                reason="PROVISIONING_ERROR",
-                message=message,
-                last_transition_at=datetime.now(timezone.utc),
-            )
-
-    def _cleanup_failed_containers(self, sandbox_id: str) -> None:
-        """
-        Best-effort cleanup for containers left behind after a failed provision.
-        """
-        label_selector = f"{SANDBOX_ID_LABEL}={sandbox_id}"
-        try:
-            containers = self.docker_client.containers.list(
-                all=True, filters={"label": label_selector}
-            )
-        except DockerException as exc:
-            logger.warning("sandbox=%s | cleanup listing failed containers: %s", sandbox_id, exc)
-            self._cleanup_egress_sidecar(sandbox_id)
-            return
-
-        for container in containers:
-            labels = container.attrs.get("Config", {}).get("Labels") or {}
-            mount_keys_raw = labels.get(SANDBOX_OSSFS_MOUNTS_LABEL, "[]")
-            try:
-                mount_keys: list[str] = json.loads(mount_keys_raw)
-            except (TypeError, json.JSONDecodeError):
-                mount_keys = []
-            try:
-                with self._docker_operation("cleanup failed sandbox container", sandbox_id):
-                    container.remove(force=True)
-            except DockerException as exc:
-                logger.warning(
-                    "sandbox=%s | failed to remove leftover container %s: %s",
-                    sandbox_id,
-                    container.id,
-                    exc,
-                )
-            finally:
-                self._release_ossfs_mounts(mount_keys)
-        # Always attempt to cleanup sidecar as well
-        self._cleanup_egress_sidecar(sandbox_id)
-
-    def _remove_pending_sandbox(self, sandbox_id: str) -> None:
-        with self._pending_lock:
-            timer = self._pending_cleanup_timers.pop(sandbox_id, None)
-            if timer:
-                timer.cancel()
-            self._pending_sandboxes.pop(sandbox_id, None)
-
-    def _get_pending_sandbox(self, sandbox_id: str) -> Optional[PendingSandbox]:
-        with self._pending_lock:
-            pending = self._pending_sandboxes.get(sandbox_id)
-            return pending
-
-    def _iter_pending_sandboxes(self) -> list[tuple[str, PendingSandbox]]:
-        with self._pending_lock:
-            return list(self._pending_sandboxes.items())
-
-    @staticmethod
-    def _pending_to_sandbox(sandbox_id: str, pending: PendingSandbox) -> Sandbox:
-        snapshot_id = getattr(pending.request, "snapshot_id", None)
-        if not isinstance(snapshot_id, str) or not snapshot_id:
-            snapshot_id = None
-        return Sandbox(
-            id=sandbox_id,
-            image=None if snapshot_id else pending.request.image,
-            snapshotId=snapshot_id,
-            platform=pending.request.platform,
-            status=pending.status,
-            metadata=pending.request.metadata,
-            entrypoint=pending.request.entrypoint,
-            expiresAt=pending.expires_at,
-            createdAt=pending.created_at,
-        )
-
-    def _update_container_labels(self, container, labels: Dict[str, str]) -> None:
-        """
-        Update container labels, falling back to raw API if docker-py lacks support.
-        """
-        try:
-            container.update(labels=labels)
-        except TypeError:
-            # Older docker-py versions do not accept labels; call low-level API directly.
-            url = self.docker_client.api._url(f"/containers/{container.id}/update")  # noqa: SLF001
-            data = {"Labels": labels}
-            self.docker_client.api._post_json(url, data=data)  # noqa: SLF001
-        container.reload()
-
-    def _schedule_pending_cleanup(self, sandbox_id: str) -> None:
-        def _cleanup():
-            self._remove_pending_sandbox(sandbox_id)
-
-        timer = Timer(PENDING_FAILURE_TTL_SECONDS, _cleanup)
-        timer.daemon = True
-        with self._pending_lock:
-            existing = self._pending_cleanup_timers.pop(sandbox_id, None)
-            if existing:
-                existing.cancel()
-            self._pending_cleanup_timers[sandbox_id] = timer
-        timer.start()
-
-    def _pull_image(
-        self,
-        image_uri: str,
-        auth_config: Optional[dict],
-        sandbox_id: str,
-        platform: Optional[PlatformSpec] = None,
-    ) -> None:
-        docker_platform = resolve_docker_platform(platform)
-        try:
-            with self._docker_operation(f"pull image {image_uri}", sandbox_id):
-                pull_kwargs: dict[str, Any] = {"auth_config": auth_config}
-                if docker_platform is not None:
-                    pull_kwargs["platform"] = docker_platform
-                self.docker_client.images.pull(image_uri, **pull_kwargs)
-        except TypeError as exc:
-            if docker_platform is not None:
-                raise HTTPException(
-                    status_code=status.HTTP_400_BAD_REQUEST,
-                    detail={
-                        "code": SandboxErrorCodes.INVALID_PARAMETER,
-                        "message": (
-                            "The configured Docker client/daemon does not support "
-                            f"platform-aware image pull for '{docker_platform}'."
-                        ),
-                    },
-                ) from exc
-            raise
-        except DockerException as exc:
-            raise HTTPException(
-                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                detail={
-                    "code": SandboxErrorCodes.IMAGE_PULL_FAILED,
-                    "message": f"Failed to pull image {image_uri}: {str(exc)}",
-                },
-            ) from exc
-
-    def _ensure_image_available(
-        self,
-        image_uri: str,
-        auth_config: Optional[dict],
-        sandbox_id: str,
-        platform: Optional[PlatformSpec] = None,
-    ) -> None:
-        expected_platform = platform
-        if expected_platform is not None and is_windows_platform(expected_platform):
-            expected_platform = None
-        if expected_platform is None:
-            expected_platform = self._get_daemon_platform()
-        try:
-            with self._docker_operation(f"inspect image {image_uri}", sandbox_id):
-                image = self.docker_client.images.get(image_uri)
-                if expected_platform is None:
-                    logger.debug(
-                        "Sandbox %s using cached image %s without platform check (daemon platform unavailable)",
-                        sandbox_id,
-                        image_uri,
-                    )
-                    return
-                image_attrs = getattr(image, "attrs", {}) or {}
-                image_os = (image_attrs.get("Os") or image_attrs.get("os") or "").lower()
-                image_arch = (
-                    image_attrs.get("Architecture")
-                    or image_attrs.get("architecture")
-                    or ""
-                ).lower()
-                image_os = self._normalize_os(image_os) or image_os
-                image_arch = self._normalize_arch(image_arch) or image_arch
-                requested_os = self._normalize_os(expected_platform.os) or expected_platform.os.lower()
-                requested_arch = (
-                    self._normalize_arch(expected_platform.arch) or expected_platform.arch.lower()
-                )
-                if image_os != requested_os or image_arch != requested_arch:
-                    logger.info(
-                        "Sandbox %s cached image %s platform mismatch (cached=%s/%s, requested=%s/%s); repulling",
-                        sandbox_id,
-                        image_uri,
-                        image_os or "unknown",
-                        image_arch or "unknown",
-                        requested_os,
-                        requested_arch,
-                    )
-                    self._pull_image(
-                        image_uri,
-                        auth_config,
-                        sandbox_id,
-                        expected_platform,
-                    )
-                    return
-                logger.debug("Sandbox %s using cached image %s", sandbox_id, image_uri)
-        except ImageNotFound:
-            self._pull_image(
-                image_uri,
-                auth_config,
-                sandbox_id,
-                platform,
-            )
-        except DockerException as exc:
-            raise HTTPException(
-                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                detail={
-                    "code": SandboxErrorCodes.IMAGE_PULL_FAILED,
-                    "message": f"Failed to inspect image {image_uri}: {str(exc)}",
-                },
-            ) from exc
-
-    def _provision_sandbox(
-        self,
-        sandbox_id: str,
-        request: CreateSandboxRequest,
-        created_at: datetime,
-        expires_at: Optional[datetime],
-        pvc_inspect_cache: Optional[dict[str, dict]] = None,
-        auto_created_volumes: Optional[list[str]] = None,
-    ) -> CreateSandboxResponse:
-        labels, environment = self._build_labels_and_env(sandbox_id, request, expires_at)
-        if auto_created_volumes:
-            labels[SANDBOX_MANAGED_VOLUMES_LABEL] = json.dumps(
-                auto_created_volumes, separators=(",", ":"),
-            )
-        image_uri, auth_config = self._resolve_image_auth(request, sandbox_id)
-        mem_limit, nano_cpus, gpu_count = self._resolve_resource_limits(request)
-        egress_token: Optional[str] = None
-        requested_windows_profile = is_windows_platform(request.platform)
-
-        if requested_windows_profile:
-            validate_windows_resource_limits(request.resource_limits.root or {})
-            validate_windows_runtime_prerequisites()
-
-        # Prepare OSSFS mounts first so binds can reference mounted host paths.
-        ossfs_mount_keys = self._prepare_ossfs_mounts(request.volumes)
-        if ossfs_mount_keys:
-            labels[SANDBOX_OSSFS_MOUNTS_LABEL] = json.dumps(
-                ossfs_mount_keys,
-                separators=(",", ":"),
-            )
-
-        sidecar_container = None
-        try:
-            # For dockur/windows profile, resourceLimits are translated to
-            # guest envs (RAM_SIZE/CPU_CORES/DISK_SIZE). Avoid applying
-            # container cgroup memory/cpu limits to the outer Linux container,
-            # which can OOM-kill QEMU during installation/runtime. GPU
-            # passthrough is likewise suppressed: the Windows guest runs inside
-            # QEMU and would not see GPUs exposed to the outer container.
-            effective_mem_limit = None if requested_windows_profile else mem_limit
-            effective_nano_cpus = None if requested_windows_profile else nano_cpus
-            effective_gpu_count = None if requested_windows_profile else gpu_count
-
-            # Build volume bind mounts from request volumes.
-            # pvc_inspect_cache carries Docker volume inspect data from the
-            # validation phase, avoiding a redundant API call.
-            volume_binds = self._build_volume_binds(request.volumes, pvc_inspect_cache)
-
-            host_config_kwargs: Dict[str, Any]
-            exposed_ports: Optional[list[str]] = ["44772", "8080"]
-            if requested_windows_profile:
-                # dockur/windows exposes RDP and noVNC/web UI on these ports.
-                # https://github.com/dockur/windows/blob/master/Dockerfile
-                exposed_ports.extend(["3389/tcp", "3389/udp", "8006/tcp"])
-            container_exposed_ports: Optional[list[str]] = exposed_ports
-
-            if request.network_policy:
-                egress_token = generate_egress_token()
-                labels[SANDBOX_EGRESS_AUTH_TOKEN_METADATA_KEY] = egress_token
-                sidecar_port_bindings = allocate_port_bindings(exposed_ports)
-                host_execd_port = sidecar_port_bindings["44772"][1]
-                host_http_port = sidecar_port_bindings["8080"][1]
-                extra_sidecar_port_bindings = {
-                    port: binding
-                    for port, binding in sidecar_port_bindings.items()
-                    if port not in {"44772", "8080"}
-                }
-                sidecar_container = self._start_egress_sidecar(
-                    sandbox_id=sandbox_id,
-                    network_policy=request.network_policy,
-                    egress_token=egress_token,
-                    host_execd_port=host_execd_port,
-                    host_http_port=host_http_port,
-                    extra_port_bindings=extra_sidecar_port_bindings,
-                )
-                labels[SANDBOX_EMBEDDING_PROXY_PORT_LABEL] = str(host_execd_port)
-                labels[SANDBOX_HTTP_PORT_LABEL] = str(host_http_port)
-                host_config_kwargs = self._base_host_config_kwargs(
-                    effective_mem_limit, effective_nano_cpus, f"container:{sidecar_container.id}",
-                    gpu_count=effective_gpu_count,
-                )
-                # Container network namespace is shared with sidecar. Docker rejects
-                # exposing ports on the main container in "container:<id>" mode.
-                container_exposed_ports = None
-                # Drop NET_ADMIN for the main container; only the sidecar should keep it
-                cap_drop = set(host_config_kwargs.get("cap_drop") or [])
-                cap_drop.add("NET_ADMIN")
-                if cap_drop:
-                    host_config_kwargs["cap_drop"] = list(cap_drop)
-            else:
-                host_config_kwargs = self._base_host_config_kwargs(
-                    effective_mem_limit, effective_nano_cpus, self.network_mode,
-                    gpu_count=effective_gpu_count,
-                )
-                if self.network_mode != HOST_NETWORK_MODE:
-                    port_bindings = allocate_port_bindings(exposed_ports)
-                    host_execd_port = port_bindings["44772"][1]
-                    host_http_port = port_bindings["8080"][1]
-                    host_config_kwargs["port_bindings"] = normalize_port_bindings(port_bindings)
-                    labels[SANDBOX_EMBEDDING_PROXY_PORT_LABEL] = str(host_execd_port)
-                    labels[SANDBOX_HTTP_PORT_LABEL] = str(host_http_port)
-                else:
-                    exposed_ports = None
-
-            # Inject volume bind mounts into Docker host config
-            if volume_binds:
-                host_config_kwargs["binds"] = volume_binds
-            if requested_windows_profile:
-                host_config_kwargs = apply_windows_runtime_host_config_defaults(
-                    host_config_kwargs,
-                    sandbox_id,
-                )
-                environment = inject_windows_resource_limits_env(
-                    environment,
-                    request.resource_limits.root or {},
-                )
-                environment = inject_windows_user_ports(environment, exposed_ports)
-
-            created_container = self._create_and_start_container(
-                sandbox_id,
-                image_uri,
-                request.entrypoint,
-                labels,
-                environment,
-                host_config_kwargs,
-                container_exposed_ports,
-                request.platform,
-            )
-        except Exception:
-            if sidecar_container is not None:
-                try:
-                    sidecar_container.remove(force=True)
-                except DockerException as cleanup_exc:
-                    logger.warning(
-                        "Failed to cleanup egress sidecar for sandbox %s: %s",
-                        sandbox_id,
-                        cleanup_exc,
-                    )
-            self._release_ossfs_mounts(ossfs_mount_keys)
-            self._cleanup_managed_volumes(sandbox_id, auto_created_volumes or [])
-            raise
-
-        status_info = SandboxStatus(
-            state="Running",
-            reason="CONTAINER_RUNNING",
-            message="Sandbox container started successfully.",
-            last_transition_at=created_at,
-        )
-
-        if expires_at is not None:
-            self._schedule_expiration(sandbox_id, expires_at)
-
-        effective_platform = self._resolve_platform_for_container(created_container, labels)
-        return CreateSandboxResponse(
-            id=sandbox_id,
-            status=status_info,
-            metadata=request.metadata,
-            platform=effective_platform or request.platform,
-            expiresAt=expires_at,
-            createdAt=created_at,
-            entrypoint=request.entrypoint,
-        )
-
-    def _is_user_defined_network(self) -> bool:
-        """Return True when network_mode is a named user-defined network (not host/bridge/none/container:*)."""
-        return (
-            self.network_mode not in {HOST_NETWORK_MODE, BRIDGE_NETWORK_MODE, "none"}
-            and not self.network_mode.startswith("container:")
-        )
-
-    def _validate_network_exists(self) -> None:
-        """Verify the configured user-defined Docker network exists before creating a sandbox."""
-        if not self._is_user_defined_network():
-            return
-        try:
-            self.docker_client.networks.get(self.network_mode)
-        except DockerNotFound:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail={
-                    "code": SandboxErrorCodes.INVALID_PARAMETER,
-                    "message": (
-                        f"Docker network '{self.network_mode}' does not exist. "
-                        "Create it first with 'docker network create <name>'."
-                    ),
-                },
-            )
-        except DockerException as exc:
-            raise HTTPException(
-                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                detail={
-                    "code": SandboxErrorCodes.CONTAINER_START_FAILED,
-                    "message": f"Failed to inspect Docker network '{self.network_mode}': {exc}",
-                },
-            ) from exc
-
-    def _ensure_network_policy_support(self, request: CreateSandboxRequest) -> None:
-        """
-        Validate that network policy can be honored under the current runtime config.
-
-        This includes Docker-specific checks (network_mode) and common checks (egress.image).
-        """
-        if not request.network_policy:
-            return
-
-        # Docker-specific validation: network_mode must be bridge
-        if self.network_mode == HOST_NETWORK_MODE:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail={
-                    "code": SandboxErrorCodes.INVALID_PARAMETER,
-                    "message": "networkPolicy is not supported when docker network_mode=host.",
-                },
-            )
-
-        # User-defined networks cannot be combined with networkPolicy: the egress sidecar
-        # always runs on the default bridge, which would silently discard the configured network.
-        if self._is_user_defined_network():
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail={
-                    "code": SandboxErrorCodes.INVALID_PARAMETER,
-                    "message": (
-                        f"networkPolicy is not supported when docker network_mode='{self.network_mode}' "
-                        "(user-defined network). Use network_mode='bridge' to enable network policy enforcement."
-                    ),
-                },
-            )
-
-        # Common validation: egress.image must be configured
-        ensure_egress_configured(request.network_policy, self.app_config.egress)
-
-    def _ensure_secure_access_support(self, request: CreateSandboxRequest) -> None:
-        """Validate that secure access can be honored under the current Docker runtime."""
-        if not request.secure_access:
-            return
-
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail={
-                "code": SandboxErrorCodes.INVALID_PARAMETER,
-                "message": (
-                    "secureAccess is not supported when runtime.type='docker'. "
-                    "Use the Kubernetes runtime to create secured sandboxes."
-                ),
-            },
-        )
-
-    def _validate_volumes(
-        self, request: CreateSandboxRequest
-    ) -> tuple[dict[str, dict], list[str]]:
-        """
-        Validate volume definitions for Docker runtime.
-
-        Performs comprehensive validation:
-        - Calls shared volume validation (name, mount path, sub path, backend count)
-        - Delegates to backend-specific validators for Docker-level checks
-
-        Args:
-            request: Sandbox creation request.
-
-        Returns:
-            A tuple of:
-            - A dict mapping PVC volume names (``pvc.claimName``) to their
-              ``docker volume inspect`` results.  Empty when there are no PVC
-              volumes.  This data is passed to ``_build_volume_binds`` so that
-              bind generation does not need a second API call.
-            - A list of Docker named volume names that were auto-created during
-              validation (empty when ``createIfNotExists`` is false or all
-              volumes already existed).
-
-        Raises:
-            HTTPException: When any validation fails.
-        """
-        if not request.volumes:
-            return {}, []
-
-        # Shared validation: names, mount paths, sub paths, backend count, host path allowlist
-        allowed_prefixes = self.app_config.storage.allowed_host_paths
-        ensure_volumes_valid(request.volumes, allowed_host_prefixes=allowed_prefixes)
-
-        pvc_inspect_cache: dict[str, dict] = {}
-        auto_created_volumes: list[str] = []
-        try:
-            for volume in request.volumes:
-                if volume.host is not None:
-                    self._validate_host_volume(volume, allowed_prefixes)
-                elif volume.pvc is not None:
-                    vol_info, was_created = self._validate_pvc_volume(volume)
-                    pvc_inspect_cache[volume.pvc.claim_name] = vol_info
-                    if was_created and volume.pvc.delete_on_sandbox_termination:
-                        auto_created_volumes.append(volume.pvc.claim_name)
-                elif volume.ossfs is not None:
-                    self._validate_ossfs_volume(volume)
-        except Exception:
-            # If any subsequent volume validation fails, remove volumes we
-            # already auto-created so they don't leak — delete_sandbox will
-            # never run for a sandbox that was never provisioned.
-            self._cleanup_managed_volumes("<pre-sandbox>", auto_created_volumes)
-            raise
-
-        return pvc_inspect_cache, auto_created_volumes
-
-    @staticmethod
-    def _validate_host_volume(volume, allowed_prefixes: Optional[list[str]]) -> None:
-        """
-        Docker-specific validation for host bind mount volumes.
-
-        Validates that the resolved host path (host.path + optional subPath)
-        remains within allowed prefixes — including symlink resolution — then
-        ensures the directory exists on the filesystem, creating it automatically
-        if it does not.
-
-        Args:
-            volume: Volume with host backend.
-            allowed_prefixes: Optional allowlist of host path prefixes.
-
-        Raises:
-            HTTPException: When the resolved path is invalid or cannot be created.
-        """
-        resolved_path = volume.host.path
-        if volume.sub_path:
-            resolved_path = os.path.normpath(os.path.join(resolved_path, volume.sub_path))
-
-        # Defense in depth: re-validate the resolved path against the
-        # allowlist.  Even though sub_path traversal (../) is blocked by
-        # ensure_valid_sub_path(), normalizing and re-checking prevents
-        # any edge-case bypass.
-        if allowed_prefixes and resolved_path != volume.host.path:
-            ensure_valid_host_path(resolved_path, allowed_prefixes)
-
-        # ── Symlink-aware allowlist check ──
-        # os.path.normpath and ensure_valid_host_path only perform lexical
-        # checks.  A symlink within a whitelisted directory (e.g.
-        # /data/opensandbox/link -> /) would pass lexical validation but
-        # Docker resolves the symlink when creating the bind mount, escaping
-        # the allowed prefix.  Resolve both the host path and the allowed
-        # prefixes with realpath to detect this.
-        if allowed_prefixes:
-            canonical = os.path.realpath(resolved_path)
-            canonical_prefixes = [os.path.realpath(p) for p in allowed_prefixes]
-            if canonical != resolved_path or canonical_prefixes != allowed_prefixes:
-                ensure_valid_host_path(canonical, canonical_prefixes)
-
-        # Allow existing host files (for example ISO binds to /boot.iso)
-        # without attempting directory creation.
-        if os.path.isfile(resolved_path):
-            return
-
-        try:
-            os.makedirs(resolved_path, exist_ok=True)
-        except OSError as e:
-            raise HTTPException(
-                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                detail={
-                    "code": SandboxErrorCodes.HOST_PATH_CREATE_FAILED,
-                    "message": (
-                        f"Volume '{volume.name}': could not ensure host path "
-                        f"directory exists at '{resolved_path}': {type(e).__name__}"
-                    ),
-                },
-            )
-
-    def _validate_pvc_volume(self, volume) -> tuple[dict, bool]:
-        """
-        Docker-specific validation for PVC (named volume) backend.
-
-        In Docker runtime, the ``pvc`` backend maps to a Docker named volume.
-        ``pvc.claimName`` is used as the Docker volume name.  The volume must
-        already exist (created via ``docker volume create``).
-
-        When ``subPath`` is specified, the volume must use the ``local`` driver
-        so that the host-side ``Mountpoint`` is a real filesystem path.  The
-        resolved path (``Mountpoint + subPath``) is validated for path-traversal
-        safety but *not* for existence, because the Mountpoint directory is
-        typically owned by root and may not be stat-able by the server process.
-
-        Args:
-            volume: Volume with pvc backend.
-
-        Returns:
-            A tuple of:
-            - The ``docker volume inspect`` result dict for the named volume.
-            - Whether the volume was auto-created by this call.
-
-        Raises:
-            HTTPException: When the named volume does not exist, inspection
-                fails, or subPath constraints are violated.
-        """
-        volume_name = volume.pvc.claim_name
-        auto_created = False
-        try:
-            vol_info = self.docker_client.api.inspect_volume(volume_name)
-        except DockerNotFound:
-            if volume.pvc.create_if_not_exists:
-                # Auto-create the Docker named volume
-                try:
-                    self.docker_client.api.create_volume(
-                        name=volume_name,
-                        labels={SANDBOX_MANAGED_VOLUMES_LABEL: "server"},
-                    )
-                    logger.info("Auto-created Docker named volume '%s'", volume_name)
-                    vol_info = self.docker_client.api.inspect_volume(volume_name)
-                    auto_created = True
-                except DockerException as create_exc:
-                    raise HTTPException(
-                        status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                        detail={
-                            "code": SandboxErrorCodes.PVC_VOLUME_INSPECT_FAILED,
-                            "message": (
-                                f"Volume '{volume.name}': failed to auto-create Docker "
-                                f"named volume '{volume_name}': {create_exc}"
-                            ),
-                        },
-                    ) from create_exc
-            else:
-                raise HTTPException(
-                    status_code=status.HTTP_400_BAD_REQUEST,
-                    detail={
-                        "code": SandboxErrorCodes.PVC_VOLUME_NOT_FOUND,
-                        "message": (
-                            f"Volume '{volume.name}': Docker named volume '{volume_name}' "
-                            "does not exist. Named volumes must be created before sandbox "
-                            "creation (e.g., 'docker volume create <name>')."
-                        ),
-                    },
-                )
-        except DockerException as exc:
-            raise HTTPException(
-                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                detail={
-                    "code": SandboxErrorCodes.PVC_VOLUME_INSPECT_FAILED,
-                    "message": (
-                        f"Volume '{volume.name}': failed to inspect Docker named volume "
-                        f"'{volume_name}': {exc}"
-                    ),
-                },
-            ) from exc
-
-        # --- subPath validation for Docker named volumes ---
-        if volume.sub_path:
-            driver = vol_info.get("Driver", "")
-            if driver != "local":
-                raise HTTPException(
-                    status_code=status.HTTP_400_BAD_REQUEST,
-                    detail={
-                        "code": SandboxErrorCodes.PVC_SUBPATH_UNSUPPORTED_DRIVER,
-                        "message": (
-                            f"Volume '{volume.name}': subPath is only supported for "
-                            f"Docker named volumes using the 'local' driver, but "
-                            f"volume '{volume_name}' uses driver '{driver}'."
-                        ),
-                    },
-                )
-
-            mountpoint = vol_info.get("Mountpoint", "")
-            if not mountpoint:
-                raise HTTPException(
-                    status_code=status.HTTP_400_BAD_REQUEST,
-                    detail={
-                        "code": SandboxErrorCodes.PVC_SUBPATH_UNSUPPORTED_DRIVER,
-                        "message": (
-                            f"Volume '{volume.name}': cannot resolve subPath because "
-                            f"Docker named volume '{volume_name}' has no Mountpoint."
-                        ),
-                    },
-                )
-
-            resolved_path = posixpath.normpath(
-                posixpath.join(mountpoint, volume.sub_path)
-            )
-
-            # ── Path-escape check (lexical + symlink) ──
-            #
-            # 1. Lexical check via normpath + path-boundary-aware startswith.
-            #    Use mountpoint + "/" to avoid false positives when one
-            #    mountpoint is a prefix of another (e.g., …/_data vs …/_data2).
-            #    Docker Mountpoint paths are always POSIX, so use "/" directly.
-            mountpoint_prefix = (
-                mountpoint if mountpoint.endswith("/") else mountpoint + "/"
-            )
-            if resolved_path != mountpoint and not resolved_path.startswith(
-                mountpoint_prefix
-            ):
-                raise HTTPException(
-                    status_code=status.HTTP_400_BAD_REQUEST,
-                    detail={
-                        "code": SandboxErrorCodes.INVALID_SUB_PATH,
-                        "message": (
-                            f"Volume '{volume.name}': resolved subPath escapes the "
-                            f"volume mountpoint."
-                        ),
-                    },
-                )
-
-            # 2. Symlink-aware check (best-effort).
-            #    Docker volume Mountpoint dirs are typically root-owned and not
-            #    readable by the server process.  Using strict=True so that
-            #    realpath raises OSError when it cannot traverse a directory
-            #    instead of silently returning the unresolved lexical path
-            #    (which would make this check a no-op).  When the path IS
-            #    accessible, this detects symlink-escape attacks (e.g., a
-            #    malicious symlink datasets -> /).
-            try:
-                canonical_mountpoint = os.path.realpath(
-                    mountpoint, strict=True
-                )
-                canonical_resolved = os.path.realpath(
-                    resolved_path, strict=True
-                )
-                # os.path.realpath returns OS-native separators, so use
-                # os.sep here (unlike the lexical check above which operates
-                # on POSIX-normalised Docker Mountpoint strings).
-                canonical_prefix = (
-                    canonical_mountpoint
-                    if canonical_mountpoint.endswith(os.sep)
-                    else canonical_mountpoint + os.sep
-                )
-                if (
-                    canonical_resolved != canonical_mountpoint
-                    and not canonical_resolved.startswith(canonical_prefix)
-                ):
-                    raise HTTPException(
-                        status_code=status.HTTP_400_BAD_REQUEST,
-                        detail={
-                            "code": SandboxErrorCodes.INVALID_SUB_PATH,
-                            "message": (
-                                f"Volume '{volume.name}': resolved subPath escapes "
-                                f"the volume mountpoint after symlink resolution."
-                            ),
-                        },
-                    )
-            except OSError:
-                # Cannot access volume paths (expected for non-root server).
-                # Lexical validation above is still enforced; the symlink
-                # check is skipped because we cannot resolve the real paths.
-                pass
-
-            # NOTE: We intentionally do NOT check os.path.exists(resolved_path)
-            # here.  Docker volume Mountpoint directories (e.g.,
-            # /var/lib/docker/volumes/…/_data) are typically owned by root and
-            # not readable by the server process.  os.path.exists() returns
-            # False when the process lacks permission to stat the path, causing
-            # false-negative rejections.  If the subPath does not actually
-            # exist, Docker will report the error at container creation time.
-
-        return vol_info, auto_created
-
-    def _build_volume_binds(
-        self,
-        volumes: Optional[list],
-        pvc_inspect_cache: Optional[dict[str, dict]] = None,
-    ) -> list[str]:
-        """
-        Convert Volume definitions into Docker bind/volume mount specs.
-
-        Supported backends:
-        - ``host``: host path bind mount.
-          Format: ``/host/path:/container/path:ro|rw``
-        - ``pvc``: Docker named volume mount.
-          Format (no subPath): ``volume-name:/container/path:ro|rw``
-          Docker recognises non-absolute-path sources as named volume references.
-          Format (with subPath): ``/var/lib/docker/volumes/…/subdir:/container/path:ro|rw``
-          When subPath is specified, the volume's host Mountpoint (obtained from
-          ``pvc_inspect_cache``) is used to produce a standard bind mount.
-        - ``ossfs``: host bind mount to runtime-mounted OSSFS path.
-          Format: ``/mnt/ossfs/<bucket>/<subPath?>:/container/path:ro|rw``
-
-        Each mount string uses ``:ro`` for read-only and ``:rw`` for read-write
-        (default).
-
-        Args:
-            volumes: List of Volume objects from the creation request.
-            pvc_inspect_cache: Dict mapping PVC claimNames to their
-                ``docker volume inspect`` results, populated by
-                ``_validate_volumes``.  Avoids a redundant API call and
-                eliminates the race window between validation and bind
-                generation.
-
-        Returns:
-            List of Docker bind/volume mount strings.
-        """
-        if not volumes:
-            return []
-
-        cache = pvc_inspect_cache or {}
-        binds: list[str] = []
-        for volume in volumes:
-            container_path = volume.mount_path
-            mode = "ro" if volume.read_only else "rw"
-
-            if volume.host is not None:
-                # Resolve the concrete host path (host.path + optional subPath)
-                host_path = volume.host.path
-                if volume.sub_path:
-                    host_path = os.path.normpath(
-                        os.path.join(host_path, volume.sub_path)
-                    )
-                binds.append(f"{host_path}:{container_path}:{mode}")
-
-            elif volume.pvc is not None:
-                if volume.sub_path:
-                    # Resolve the named volume's host-side Mountpoint and append
-                    # the subPath to produce a regular bind mount.  Validation
-                    # has already ensured the driver is "local" and the resolved
-                    # path is safe.  Reuse cached inspect data to avoid a
-                    # redundant Docker API call and potential race condition.
-                    vol_info = cache.get(volume.pvc.claim_name, {})
-                    mountpoint = vol_info.get("Mountpoint", "")
-                    resolved = posixpath.normpath(
-                        posixpath.join(mountpoint, volume.sub_path)
-                    )
-                    binds.append(f"{resolved}:{container_path}:{mode}")
-                else:
-                    # No subPath: use claimName directly as Docker volume ref.
-                    binds.append(
-                        f"{volume.pvc.claim_name}:{container_path}:{mode}"
-                    )
-            elif volume.ossfs is not None:
-                _, host_path = self._resolve_ossfs_paths(volume)
-                binds.append(f"{host_path}:{container_path}:{mode}")
-
-        return binds
-
-    def list_sandboxes(self, request: ListSandboxesRequest) -> ListSandboxesResponse:
-        """
-        List sandboxes with optional filtering and pagination.
-        """
-        try:
-            containers = self.docker_client.containers.list(
-                all=True,
-                filters={"label": [SANDBOX_ID_LABEL]},
-            )
-        except DockerException as exc:
-            raise HTTPException(
-                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                detail={
-                    "code": SandboxErrorCodes.CONTAINER_QUERY_FAILED,
-                    "message": f"Failed to query sandbox containers: {str(exc)}",
-                },
-            ) from exc
-
-        sandboxes_by_id: dict[str, Sandbox] = {}
-        container_ids: set[str] = set()
-        for container in containers:
-            labels = container.attrs.get("Config", {}).get("Labels") or {}
-            sandbox_id = labels.get(SANDBOX_ID_LABEL)
-            if not sandbox_id:
-                continue
-            sandbox_obj = self._container_to_sandbox(container, sandbox_id)
-            container_ids.add(sandbox_id)
-            if matches_filter(sandbox_obj, request.filter):
-                sandboxes_by_id[sandbox_id] = sandbox_obj
-
-        for sandbox_id, pending in self._iter_pending_sandboxes():
-            if sandbox_id in container_ids:
-                # If a real container exists, prefer its state regardless of filter outcome.
-                continue
-            sandbox_obj = self._pending_to_sandbox(sandbox_id, pending)
-            if matches_filter(sandbox_obj, request.filter):
-                sandboxes_by_id[sandbox_id] = sandbox_obj
-
-        sandboxes: list[Sandbox] = list(sandboxes_by_id.values())
-
-        sandboxes.sort(key=lambda s: s.created_at or datetime.min, reverse=True)
-
-        if request.pagination:
-            page = request.pagination.page
-            page_size = request.pagination.page_size
-        else:
-            page = 1
-            page_size = 20
-
-        total_items = len(sandboxes)
-        total_pages = math.ceil(total_items / page_size) if total_items else 0
-        start_index = (page - 1) * page_size
-        end_index = start_index + page_size
-        items = sandboxes[start_index:end_index]
-        has_next_page = page < total_pages
-
-        pagination_info = PaginationInfo(
-            page=page,
-            page_size=page_size,
-            total_items=total_items,
-            total_pages=total_pages,
-            has_next_page=has_next_page,
-        )
-
-        return ListSandboxesResponse(items=items, pagination=pagination_info)
-
-    def get_sandbox(self, sandbox_id: str) -> Sandbox:
-        """
-        Fetch a sandbox by id.
-
-        Args:
-            sandbox_id: Unique sandbox identifier
-
-        Returns:
-            Sandbox: Complete sandbox information
-
-        Raises:
-            HTTPException: If sandbox not found
-        """
-        # Prefer real container state; fall back to pending record only if no container exists.
-        try:
-            container = self._get_container_by_sandbox_id(sandbox_id)
-        except HTTPException as exc:
-            if exc.status_code != status.HTTP_404_NOT_FOUND:
-                raise
-            pending = self._get_pending_sandbox(sandbox_id)
-            if pending:
-                return self._pending_to_sandbox(sandbox_id, pending)
-            raise
-        return self._container_to_sandbox(container, sandbox_id)
-
-    def delete_sandbox(self, sandbox_id: str) -> None:
-        """
-        Delete a sandbox using Docker.
-
-        Args:
-            sandbox_id: Unique sandbox identifier
-
-        Raises:
-            HTTPException: If sandbox not found or deletion fails
-        """
-        container = self._get_container_by_sandbox_id(sandbox_id)
-        labels = container.attrs.get("Config", {}).get("Labels") or {}
-        mount_keys_raw = labels.get(SANDBOX_OSSFS_MOUNTS_LABEL, "[]")
-        try:
-            mount_keys: list[str] = json.loads(mount_keys_raw)
-        except (TypeError, json.JSONDecodeError):
-            mount_keys = []
-        managed_volumes_raw = labels.get(SANDBOX_MANAGED_VOLUMES_LABEL, "[]")
-        try:
-            managed_volumes: list[str] = json.loads(managed_volumes_raw)
-        except (TypeError, json.JSONDecodeError):
-            managed_volumes = []
-        try:
-            try:
-                with self._docker_operation("kill sandbox container", sandbox_id):
-                    container.kill()
-            except DockerException as exc:
-                # Ignore error if container is already stopped
-                if "is not running" not in str(exc).lower():
-                    raise
-            with self._docker_operation("remove sandbox container", sandbox_id):
-                container.remove(force=True)
-        except DockerException as exc:
-            raise HTTPException(
-                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                detail={
-                    "code": SandboxErrorCodes.SANDBOX_DELETE_FAILED,
-                    "message": f"Failed to delete sandbox container: {str(exc)}",
-                },
-            ) from exc
-        finally:
-            self._remove_expiration_tracking(sandbox_id)
-            self._cleanup_egress_sidecar(sandbox_id)
-            self._cleanup_windows_oem_volume(sandbox_id, labels)
-            self._release_ossfs_mounts(mount_keys)
-            self._cleanup_managed_volumes(sandbox_id, managed_volumes)
-
-    def _cleanup_windows_oem_volume(
-        self,
-        sandbox_id: str,
-        labels: Optional[dict[str, str]],
-    ) -> None:
-        """Best-effort cleanup for windows profile OEM named volume."""
-        if labels is not None and labels.get(SANDBOX_PLATFORM_OS_LABEL) != "windows":
-            return
-
-        volume_name = f"{WINDOWS_OEM_VOLUME_PREFIX}-{sandbox_id}"
-        try:
-            with self._docker_operation("remove windows oem volume", sandbox_id):
-                self.docker_client.api.remove_volume(volume_name)
-        except DockerNotFound:
-            return
-        except DockerException as exc:
-            logger.warning(
-                "sandbox=%s | failed to remove windows OEM volume %s: %s",
-                sandbox_id,
-                volume_name,
-                exc,
-            )
-
-    def pause_sandbox(self, sandbox_id: str) -> None:
-        """
-        Pause a running sandbox using Docker.
-
-        Args:
-            sandbox_id: Unique sandbox identifier
-
-        Raises:
-            HTTPException: If sandbox not found or cannot be paused
-        """
-        container = self._get_container_by_sandbox_id(sandbox_id)
-        state = container.attrs.get("State", {})
-        if not state.get("Running", False):
-            raise HTTPException(
-                status_code=status.HTTP_409_CONFLICT,
-                detail={
-                    "code": SandboxErrorCodes.SANDBOX_NOT_RUNNING,
-                    "message": "Sandbox is not in a running state.",
-                },
-            )
-
-        try:
-            with self._docker_operation("pause sandbox container", sandbox_id):
-                container.pause()
-        except DockerException as exc:
-            raise HTTPException(
-                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                detail={
-                    "code": SandboxErrorCodes.SANDBOX_PAUSE_FAILED,
-                    "message": f"Failed to pause sandbox container: {str(exc)}",
-                },
-            ) from exc
-
-    def resume_sandbox(self, sandbox_id: str) -> None:
-        """
-        Resume a paused sandbox using Docker.
-
-        Args:
-            sandbox_id: Unique sandbox identifier
-
-        Raises:
-            HTTPException: If sandbox not found or cannot be resumed
-        """
-        container = self._get_container_by_sandbox_id(sandbox_id)
-        state = container.attrs.get("State", {})
-        if not state.get("Paused", False):
-            raise HTTPException(
-                status_code=status.HTTP_409_CONFLICT,
-                detail={
-                    "code": SandboxErrorCodes.SANDBOX_NOT_PAUSED,
-                    "message": "Sandbox is not in a paused state.",
-                },
-            )
-
-        try:
-            with self._docker_operation("resume sandbox container", sandbox_id):
-                container.unpause()
-        except DockerException as exc:
-            raise HTTPException(
-                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                detail={
-                    "code": SandboxErrorCodes.SANDBOX_RESUME_FAILED,
-                    "message": f"Failed to resume sandbox container: {str(exc)}",
-                },
-            ) from exc
-
-    def get_access_renew_extend_seconds(self, sandbox_id: str) -> Optional[int]:
-        try:
-            container = self._get_container_by_sandbox_id(sandbox_id)
-        except HTTPException:
-            return None
-        labels = container.attrs.get("Config", {}).get("Labels") or {}
-        raw = labels.get(ACCESS_RENEW_EXTEND_SECONDS_METADATA_KEY)
-        if raw is None or not str(raw).strip():
-            return None
-        try:
-            return int(str(raw).strip())
-        except ValueError:
-            return None
-
-    def renew_expiration(
-        self,
-        sandbox_id: str,
-        request: RenewSandboxExpirationRequest,
-    ) -> RenewSandboxExpirationResponse:
-        """
-        Renew sandbox expiration time.
-
-        Args:
-            sandbox_id: Unique sandbox identifier
-            request: Renewal request with new expiration time
-
-        Returns:
-            RenewSandboxExpirationResponse: Updated expiration time
-
-        Raises:
-            HTTPException: If sandbox not found or renewal fails
-        """
-        container = self._get_container_by_sandbox_id(sandbox_id)
-        new_expiration = ensure_future_expiration(request.expires_at)
-
-        labels = container.attrs.get("Config", {}).get("Labels") or {}
-        if self._has_manual_cleanup(labels):
-            raise HTTPException(
-                status_code=status.HTTP_409_CONFLICT,
-                detail={
-                    "code": SandboxErrorCodes.INVALID_EXPIRATION,
-                    "message": f"Sandbox {sandbox_id} does not have automatic expiration enabled.",
-                },
-            )
-        if self._get_tracked_expiration(sandbox_id, labels) is None:
-            raise HTTPException(
-                status_code=status.HTTP_409_CONFLICT,
-                detail={
-                    "code": SandboxErrorCodes.INVALID_EXPIRATION,
-                    "message": (
-                        f"Sandbox {sandbox_id} is missing expiration metadata and cannot be renewed safely."
-                    ),
-                },
-            )
-
-        # Persist the new timeout in memory; it will also be respected on restart via _restore_existing_sandboxes
-        self._schedule_expiration(sandbox_id, new_expiration)
-        labels[SANDBOX_EXPIRES_AT_LABEL] = new_expiration.isoformat()
-        try:
-            with self._docker_operation("update sandbox labels", sandbox_id):
-                self._update_container_labels(container, labels)
-        except (DockerException, TypeError) as exc:
-            logger.warning("Failed to refresh labels for sandbox %s: %s", sandbox_id, exc)
-
-        return RenewSandboxExpirationResponse(expires_at=new_expiration)
-
-    def get_endpoint(self, sandbox_id: str, port: int, resolve_internal: bool = False,
-                     expires: Optional[int] = None) -> Endpoint:
-        """
-        Get sandbox access endpoint.
-
-        Args:
-            sandbox_id: Unique sandbox identifier
-            port: Port number where the service is listening inside the sandbox
-            resolve_internal: If True, return the internal container IP (for proxy), ignoring router config.
-            expires: Not supported by Docker runtime.
-
-        Returns:
-            Endpoint: Public endpoint URL
-
-        Raises:
-            HTTPException: If sandbox not found, endpoint not available,
-                or expires is provided (Docker does not support signed routes).
-        """
-        if expires is not None:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail={
-                    "code": SandboxErrorCodes.API_NOT_SUPPORTED,
-                    "message": (
-                        "Signed routes (expires parameter) are not supported when "
-                        "runtime.type='docker'. Use the Kubernetes runtime for signed routes."
-                    ),
-                },
-            )
-
-        try:
-            self.validate_port(port)
-        except ValueError as exc:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail={
-                    "code": SandboxErrorCodes.INVALID_PORT,
-                    "message": str(exc),
-                },
-            ) from exc
-
-        if resolve_internal:
-            container = self._get_container_by_sandbox_id(sandbox_id)
-            labels = container.attrs.get("Config", {}).get("Labels") or {}
-            # Sandboxes created with egress sidecar share the sidecar network namespace, so the
-            # main container's private IP is not a stable proxy target. In that case, treat the
-            # server-proxy target as the server-local host-mapped endpoint instead of a container IP.
-            if labels.get(SANDBOX_EGRESS_AUTH_TOKEN_METADATA_KEY):
-                return self._resolve_host_mapped_endpoint(
-                    self._resolve_proxy_host(),
-                    labels,
-                    port,
-                )
-            return self._resolve_internal_endpoint(container, port)
-
-        public_host = self._resolve_public_host()
-
-        if self.network_mode == HOST_NETWORK_MODE:
-            endpoint = Endpoint(endpoint=f"{public_host}:{port}")
-            container = self._get_container_by_sandbox_id(sandbox_id)
-            labels = container.attrs.get("Config", {}).get("Labels") or {}
-            self._attach_egress_auth_headers(endpoint, labels)
-            return endpoint
-
-        # non-host mode (bridge / user-defined network)
-        container = self._get_container_by_sandbox_id(sandbox_id)
-        labels = container.attrs.get("Config", {}).get("Labels") or {}
-        return self._resolve_host_mapped_endpoint(public_host, labels, port)
-
-    def _resolve_host_mapped_endpoint(
-        self,
-        public_host: str,
-        labels: dict[str, str],
-        port: int,
-    ) -> Endpoint:
-        execd_host_port = self._parse_host_port_label(
-            labels.get(SANDBOX_EMBEDDING_PROXY_PORT_LABEL),
-            SANDBOX_EMBEDDING_PROXY_PORT_LABEL,
-        )
-        http_host_port = self._parse_host_port_label(
-            labels.get(SANDBOX_HTTP_PORT_LABEL),
-            SANDBOX_HTTP_PORT_LABEL,
-        )
-
-        if port == 8080:
-            if http_host_port is None:
-                raise HTTPException(
-                    status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                    detail={
-                        "code": SandboxErrorCodes.NETWORK_MODE_ENDPOINT_UNAVAILABLE,
-                        "message": "Missing host port mapping for container port 8080.",
-                    },
-                )
-            endpoint = Endpoint(endpoint=f"{public_host}:{http_host_port}")
-            self._attach_egress_auth_headers(endpoint, labels)
-            return endpoint
-
-        if execd_host_port is None:
-            raise HTTPException(
-                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                detail={
-                    "code": SandboxErrorCodes.NETWORK_MODE_ENDPOINT_UNAVAILABLE,
-                    "message": "Missing host port mapping for execd proxy port 44772.",
-                },
-            )
-
-        endpoint = Endpoint(endpoint=f"{public_host}:{execd_host_port}/proxy/{port}")
-        self._attach_egress_auth_headers(endpoint, labels)
-        return endpoint
-
-    def _attach_egress_auth_headers(
-        self,
-        endpoint: Endpoint,
-        labels: dict[str, str],
-    ) -> None:
-        token = labels.get(SANDBOX_EGRESS_AUTH_TOKEN_METADATA_KEY)
-        if not token:
-            return
-        endpoint.headers = merge_endpoint_headers(
-            endpoint.headers,
-            build_egress_auth_headers(token),
-        )
-
-    def _get_docker_host_ip(self) -> Optional[str]:
-        """When running inside a container, return [docker].host_ip for endpoint URLs (if set)."""
-        ip = (self.app_config.docker.host_ip or "").strip()
-        return ip or None
-
-    def _resolve_public_host(self) -> str:
-        """Resolve the host used in endpoint URLs. If [server].eip is set, use it directly without resolving host."""
-        eip_cfg = (self.app_config.server.eip or "").strip()
-        if eip_cfg:
-            return eip_cfg
-        host_cfg = (self.app_config.server.host or "").strip()
-        host_key = host_cfg.lower()
-        if host_key in {"", "0.0.0.0", "::"}:
-            if _running_inside_docker_container():
-                host_ip = self._get_docker_host_ip()
-                if host_ip:
-                    return host_ip
-            return self._resolve_bind_ip(socket.AF_INET)
-        return host_cfg
-
-    def _resolve_proxy_host(self) -> str:
-        """Resolve the server-local host used for proxying to host-mapped Docker endpoints.
-
-        This intentionally does not use ``server.eip`` because the proxy target must be reachable
-        from the server process itself, even in deployments without hairpin access to the public EIP.
-        """
-        host_cfg = (self.app_config.server.host or "").strip()
-        host_key = host_cfg.lower()
-        if host_key in {"", "0.0.0.0", "::"}:
-            if _running_inside_docker_container():
-                host_ip = self._get_docker_host_ip()
-                if host_ip:
-                    return host_ip
-            return "127.0.0.1"
-        return host_cfg
-
-    def _resolve_internal_endpoint(self, container, port: int) -> Endpoint:
-        """Return the internal endpoint used when bypassing host mapping."""
-        if self.network_mode == HOST_NETWORK_MODE:
-            return Endpoint(endpoint=f"127.0.0.1:{port}")
-
-        ip_address = self._extract_bridge_ip(container)
-        return Endpoint(endpoint=f"{ip_address}:{port}")
-
-    # ---------------------------
-    # Common helpers for creation
-    # ---------------------------
-    def _build_labels_and_env(
-        self,
-        sandbox_id: str,
-        request: CreateSandboxRequest,
-        expires_at: Optional[datetime],
-    ) -> tuple[dict[str, str], list[str]]:
-        metadata = request.metadata or {}
-        labels = {key: str(value) for key, value in metadata.items()}
-        labels[SANDBOX_ID_LABEL] = sandbox_id
-        if expires_at is None:
-            labels[SANDBOX_MANUAL_CLEANUP_LABEL] = "true"
-        else:
-            labels[SANDBOX_EXPIRES_AT_LABEL] = expires_at.isoformat()
-        if request.platform is not None:
-            labels[SANDBOX_PLATFORM_OS_LABEL] = request.platform.os
-            labels[SANDBOX_PLATFORM_ARCH_LABEL] = request.platform.arch
-        if request.snapshot_id:
-            labels[SANDBOX_SNAPSHOT_ID_LABEL] = request.snapshot_id
-
-        apply_access_renew_extend_seconds_to_mapping(labels, request.extensions)
-
-        env_dict = request.env or {}
-        environment = []
-        for key, value in env_dict.items():
-            if value is None:
-                continue
-            environment.append(f"{key}={value}")
-        return labels, environment
-
-    def _resolve_image_auth(
-        self, request: CreateSandboxRequest, sandbox_id: str
-    ) -> tuple[str, Optional[dict]]:
-        image_uri = request.image.uri
-        auth_config = None
-        if request.image.auth:
-            auth_config = {
-                "username": request.image.auth.username,
-                "password": request.image.auth.password,
-            }
-        self._ensure_image_available(image_uri, auth_config, sandbox_id, request.platform)
-        return image_uri, auth_config
-
-    def _resolve_resource_limits(
-        self, request: CreateSandboxRequest
-    ) -> tuple[Optional[int], Optional[int], Optional[int]]:
-        resource_limits = request.resource_limits.root or {}
-        mem_limit = parse_memory_limit(resource_limits.get("memory"))
-        nano_cpus = parse_nano_cpus(resource_limits.get("cpu"))
-        gpu_count = parse_gpu_request(resource_limits.get("gpu"))
-        return mem_limit, nano_cpus, gpu_count
-
-    def _base_host_config_kwargs(
-        self,
-        mem_limit: Optional[int],
-        nano_cpus: Optional[int],
-        network_mode: str,
-        gpu_count: Optional[int] = None,
-    ) -> Dict[str, Any]:
-        host_config_kwargs: Dict[str, Any] = {"network_mode": network_mode}
-        security_opts: list[str] = []
-        docker_cfg = self.app_config.docker
-        if docker_cfg.no_new_privileges:
-            security_opts.append("no-new-privileges:true")
-        if docker_cfg.apparmor_profile:
-            security_opts.append(f"apparmor={docker_cfg.apparmor_profile}")
-        if docker_cfg.seccomp_profile:
-            security_opts.append(f"seccomp={docker_cfg.seccomp_profile}")
-        if security_opts:
-            host_config_kwargs["security_opt"] = security_opts
-        if docker_cfg.drop_capabilities:
-            host_config_kwargs["cap_drop"] = docker_cfg.drop_capabilities
-        if docker_cfg.pids_limit is not None:
-            host_config_kwargs["pids_limit"] = docker_cfg.pids_limit
-        if mem_limit:
-            host_config_kwargs["mem_limit"] = mem_limit
-        if nano_cpus:
-            host_config_kwargs["nano_cpus"] = nano_cpus
-        if gpu_count:
-            # Honors host toolchains such as nvidia-container-toolkit. The Docker
-            # Engine returns a clear error at container create time if the host
-            # cannot satisfy the request, so failure is surfaced rather than silent.
-            host_config_kwargs["device_requests"] = [
-                DeviceRequest(count=gpu_count, capabilities=[["gpu"]])
-            ]
-        # Inject secure runtime into host_config
-        if self.docker_runtime:
-            logger.info(
-                "Using Docker runtime '%s' for container creation",
-                self.docker_runtime,
-            )
-            host_config_kwargs["runtime"] = self.docker_runtime
-        return host_config_kwargs
-
-    def _cleanup_managed_volumes(self, sandbox_id: str, volume_names: list[str]) -> None:
-        """
-        Remove Docker named volumes that were auto-created for this sandbox.
-
-        Only volumes whose ``opensandbox.io/volume-managed-by`` label equals
-        ``"server"`` are removed.  Pre-existing volumes are never touched.
-        Errors are logged but do not propagate — volume cleanup is best-effort.
-        """
-        for name in volume_names:
-            try:
-                vol_info = self.docker_client.api.inspect_volume(name)
-                vol_labels = vol_info.get("Labels") or {}
-                if vol_labels.get(SANDBOX_MANAGED_VOLUMES_LABEL) != "server":
-                    logger.debug(
-                        "sandbox=%s | volume '%s' not managed by server, skipping removal",
-                        sandbox_id, name,
-                    )
-                    continue
-                self.docker_client.api.remove_volume(name)
-                logger.info("sandbox=%s | removed managed volume '%s'", sandbox_id, name)
-            except DockerNotFound:
-                logger.debug(
-                    "sandbox=%s | managed volume '%s' already removed", sandbox_id, name,
-                )
-            except DockerException as exc:
-                logger.warning(
-                    "sandbox=%s | failed to remove managed volume '%s': %s",
-                    sandbox_id, name, exc,
-                )
-
-    def _cleanup_egress_sidecar(self, sandbox_id: str) -> None:
-        """
-        Remove egress sidecar associated with sandbox_id (best effort).
-        """
-        try:
-            containers = self.docker_client.containers.list(
-                all=True, filters={"label": f"{EGRESS_SIDECAR_LABEL}={sandbox_id}"}
-            )
-        except DockerException as exc:
-            logger.warning("sandbox=%s | failed to list egress sidecar: %s", sandbox_id, exc)
-            return
-
-        for container in containers:
-            try:
-                with self._docker_operation("cleanup egress sidecar", sandbox_id):
-                    container.remove(force=True)
-            except DockerException as exc:
-                logger.warning(
-                    "sandbox=%s | failed to remove egress sidecar %s: %s",
-                    sandbox_id,
-                    container.id,
-                    exc,
-                )
-
-    def _start_egress_sidecar(
-        self,
-        sandbox_id: str,
-        network_policy: NetworkPolicy,
-        egress_token: str,
-        host_execd_port: int,
-        host_http_port: int,
-        extra_port_bindings: Optional[dict[str, tuple[str, int]]] = None,
-    ):
-        sidecar_name = f"sandbox-egress-{sandbox_id}"
-        sidecar_labels = {
-            EGRESS_SIDECAR_LABEL: sandbox_id,
-        }
-
-        # Ensure sidecar image is available before create/start.
-        egress_image = self.app_config.egress.image if self.app_config.egress else None
-        if not egress_image:
-            raise ValueError("egress.image must be configured when networkPolicy is provided.")
-        self._ensure_image_available(egress_image, None, sandbox_id)
-
-        policy_payload = json.dumps(network_policy.model_dump(by_alias=True, exclude_none=True))
-        assert self.app_config.egress is not None  # validated by ensure_egress_configured with networkPolicy
-        egress_mode = self.app_config.egress.mode
-        sidecar_env = [
-            f"{EGRESS_RULES_ENV}={policy_payload}",
-            f"{EGRESS_MODE_ENV}={egress_mode}",
-            f"{OPENSANDBOX_EGRESS_TOKEN}={egress_token}",
-        ]
-
-        sidecar_port_bindings: dict[str, tuple[str, int]] = {
-            "44772": ("0.0.0.0", host_execd_port),
-            "8080": ("0.0.0.0", host_http_port),
-        }
-        if extra_port_bindings:
-            sidecar_port_bindings.update(extra_port_bindings)
-
-        sidecar_host_config_kwargs: dict[str, Any] = {
-            "network_mode": BRIDGE_NETWORK_MODE,
-            "cap_add": ["NET_ADMIN"],
-            "port_bindings": normalize_port_bindings(sidecar_port_bindings),
-        }
-        if self.app_config.egress.disable_ipv6:
-            # Optional: disable IPv6 in the shared namespace when egress.disable_ipv6 is set.
-            sidecar_host_config_kwargs["sysctls"] = {
-                "net.ipv6.conf.all.disable_ipv6": 1,
-                "net.ipv6.conf.default.disable_ipv6": 1,
-                "net.ipv6.conf.lo.disable_ipv6": 1,
-            }
-
-        sidecar_host_config = self.docker_client.api.create_host_config(
-            **sidecar_host_config_kwargs
-        )
-
-        sidecar_container = None
-        sidecar_container_id: Optional[str] = None
-        try:
-            with self._docker_operation("create egress sidecar", sandbox_id):
-                sidecar_resp = self.docker_client.api.create_container(
-                    image=egress_image,
-                    name=sidecar_name,
-                    host_config=sidecar_host_config,
-                    labels=sidecar_labels,
-                    environment=sidecar_env,
-                    # Expose the ports that have host bindings so Docker publishes them in bridge mode.
-                    ports=[normalize_container_port_spec(p) for p in sidecar_port_bindings.keys()],
-                )
-            sidecar_container_id = sidecar_resp.get("Id")
-            if not sidecar_container_id:
-                raise HTTPException(
-                    status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                    detail={
-                        "code": SandboxErrorCodes.CONTAINER_START_FAILED,
-                        "message": "Docker did not return an egress sidecar container ID.",
-                    },
-                )
-            sidecar_container = self.docker_client.containers.get(sidecar_container_id)
-            with self._docker_operation("start egress sidecar", sandbox_id):
-                sidecar_container.start()
-            return sidecar_container
-        except Exception as exc:
-            if sidecar_container is not None:
-                try:
-                    with self._docker_operation("cleanup egress sidecar", sandbox_id):
-                        sidecar_container.remove(force=True)
-                except DockerException as cleanup_exc:
-                    logger.warning(
-                        "Failed to cleanup egress sidecar for sandbox %s: %s",
-                        sandbox_id,
-                        cleanup_exc,
-                    )
-            elif sidecar_container_id:
-                try:
-                    with self._docker_operation("cleanup egress sidecar (API)", sandbox_id):
-                        self.docker_client.api.remove_container(sidecar_container_id, force=True)
-                except DockerException as cleanup_exc:
-                    logger.warning(
-                        "Failed to cleanup egress sidecar for sandbox %s: %s",
-                        sandbox_id,
-                        cleanup_exc,
-                    )
-            if isinstance(exc, HTTPException):
-                raise exc
-            raise HTTPException(
-                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                detail={
-                    "code": SandboxErrorCodes.CONTAINER_START_FAILED,
-                    "message": "Egress sidecar container failed to start.",
-                },
-            ) from exc
-
-    def _create_and_start_container(
-        self,
-        sandbox_id: str,
-        image_uri: str,
-        bootstrap_command: list[str],
-        labels: dict[str, str],
-        environment: list[str],
-        host_config_kwargs: Dict[str, Any],
-        exposed_ports: Optional[list[str]],
-        platform: Optional[PlatformSpec],
-    ):
-        requested_windows_platform = is_windows_platform(platform)
-        bootstrap_command = normalize_bootstrap_command(
-            bootstrap_command,
-            requested_windows_platform,
-        )
-        docker_platform = resolve_docker_platform(platform)
-
-        host_config = self.docker_client.api.create_host_config(**host_config_kwargs)
-        container = None
-        container_id: Optional[str] = None
-        try:
-            with self._docker_operation("create sandbox container", sandbox_id):
-                container_kwargs = {
-                    "image": image_uri,
-                    "command": bootstrap_command,
-                    "ports": (
-                        [normalize_container_port_spec(p) for p in exposed_ports]
-                        if exposed_ports
-                        else None
-                    ),
-                    "name": f"sandbox-{sandbox_id}",
-                    "environment": environment,
-                    "labels": labels,
-                    "host_config": host_config,
-                }
-                if not requested_windows_platform:
-                    container_kwargs["entrypoint"] = [BOOTSTRAP_PATH]
-                if docker_platform is not None:
-                    container_kwargs["platform"] = docker_platform
-
-                response = self.docker_client.api.create_container(**container_kwargs)
-            container_id = response.get("Id")
-            if not container_id:
-                raise HTTPException(
-                    status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                    detail={
-                        "code": SandboxErrorCodes.CONTAINER_START_FAILED,
-                        "message": "Docker did not return a container ID.",
-                    },
-                )
-            container = self.docker_client.containers.get(container_id)
-            runtime_platform = self._resolve_platform_for_container(
-                container,
-                labels,
-                include_runtime_metadata=True,
-            )
-            if requested_windows_platform:
-                install_bat_bytes = fetch_execd_install_bat(
-                    docker_client=self.docker_client,
-                    execd_image=self.execd_image,
-                    cache=self._windows_profile_cache,
-                    cache_lock=self._execd_archive_lock,
-                    docker_operation=self._docker_operation,
-                    logger=logger,
-                )
-                execd_windows_bin_bytes = fetch_execd_windows_binary(
-                    docker_client=self.docker_client,
-                    execd_image=self.execd_image,
-                    cache=self._windows_profile_cache,
-                    cache_lock=self._execd_archive_lock,
-                    docker_operation=self._docker_operation,
-                    logger=logger,
-                )
-                install_windows_oem_scripts(
-                    container=container,
-                    sandbox_id=sandbox_id,
-                    install_bat_bytes=install_bat_bytes,
-                    execd_windows_bin_bytes=execd_windows_bin_bytes,
-                    ensure_directory=self._ensure_directory,
-                    docker_operation=self._docker_operation,
-                )
-                logger.info(
-                    "sandbox=%s | skip linux bootstrap/runtime injection for windows profile",
-                    sandbox_id,
-                )
-            else:
-                self._prepare_sandbox_runtime(container, sandbox_id, runtime_platform)
-            with self._docker_operation("start sandbox container", sandbox_id):
-                container.start()
-            return container
-        except Exception as exc:
-            if container is not None:
-                try:
-                    with self._docker_operation("cleanup sandbox container", sandbox_id):
-                        container.remove(force=True)
-                except DockerException as cleanup_exc:
-                    logger.warning(
-                        "Failed to cleanup container for sandbox %s: %s",
-                        sandbox_id,
-                        cleanup_exc,
-                    )
-            elif container_id:
-                try:
-                    with self._docker_operation("cleanup sandbox container (API)", sandbox_id):
-                        self.docker_client.api.remove_container(container_id, force=True)
-                except DockerException as cleanup_exc:
-                    logger.warning(
-                        "Failed to cleanup container for sandbox %s: %s",
-                        sandbox_id,
-                        cleanup_exc,
-                    )
-
-            if isinstance(exc, HTTPException):
-                raise exc
-            if isinstance(exc, TypeError) and docker_platform is not None:
-                raise HTTPException(
-                    status_code=status.HTTP_400_BAD_REQUEST,
-                    detail={
-                        "code": SandboxErrorCodes.INVALID_PARAMETER,
-                        "message": (
-                            "The configured Docker client/daemon does not support "
-                            f"platform-aware container create for '{docker_platform}'."
-                        ),
-                    },
-                ) from exc
-
-            raise HTTPException(
-                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                detail={
-                    "code": SandboxErrorCodes.CONTAINER_START_FAILED,
-                    "message": f"Failed to create or start container: {str(exc)}",
-                },
-            ) from exc
-
-    @staticmethod
-    def _parse_host_port_label(value: Optional[str], label_name: str) -> Optional[int]:
-        if not value:
-            return None
-        try:
-            port = int(value)
-            if port <= 0 or port > 65535:
-                raise ValueError
-            return port
-        except ValueError:
-            logger.warning("Invalid port label %s=%s", label_name, value)
-            return None
-
-    def _extract_bridge_ip(self, container) -> str:
-        """Extract the IP address assigned to a container on a bridge or user-defined network.
-
-        For user-defined networks, the top-level ``NetworkSettings.IPAddress`` is empty;
-        the IP lives under ``NetworkSettings.Networks[<network-name>].IPAddress``.
-        This method prefers the configured ``network_mode`` entry when it is a user-defined
-        network, then falls back to any non-empty entry for robustness.
-        """
-        network_settings = container.attrs.get("NetworkSettings", {}) or {}
-        networks = network_settings.get("Networks", {}) or {}
-        ip_address: Optional[str] = None
-
-        if self._is_user_defined_network():
-            # Prefer the explicit network entry for the configured named network.
-            net_conf = networks.get(self.network_mode) or {}
-            ip_address = net_conf.get("IPAddress") or None
-
-        if not ip_address:
-            # Default bridge path (or fallback): check the top-level IPAddress first.
-            ip_address = network_settings.get("IPAddress") or None
-
-        if not ip_address:
-            # Last resort: iterate all network entries and take the first populated IP.
-            for net_conf in networks.values():
-                if net_conf and net_conf.get("IPAddress"):
-                    ip_address = net_conf.get("IPAddress")
-                    break
-
-        if not ip_address:
-            raise HTTPException(
-                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                detail={
-                    "code": SandboxErrorCodes.NETWORK_MODE_ENDPOINT_UNAVAILABLE,
-                    "message": "Container is running but has no assigned IP address.",
-                },
-            )
-        return ip_address
