+++ server/tests/test_docker_service.py
@@ -132,7 +132,7 @@ def test_env_allows_empty_string_and_skips_none():
     assert all(not item.startswith("NONE=") for item in environment)
 
 @pytest.mark.asyncio
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 async def test_create_sandbox_applies_security_defaults(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -159,7 +159,7 @@ async def test_create_sandbox_applies_security_defaults(mock_docker):
         patch.object(service, "_ensure_image_available"),
         patch.object(service, "_prepare_sandbox_runtime"),
         patch(
-            "opensandbox_server.services.docker.allocate_port_bindings",
+            "opensandbox_server.services.docker.docker_service.allocate_port_bindings",
             return_value={
                 "44772": ("0.0.0.0", 40001),
                 "8080": ("0.0.0.0", 40002),
@@ -174,7 +174,7 @@ async def test_create_sandbox_applies_security_defaults(mock_docker):
     assert host_config.get("pids_limit") == service.app_config.docker.pids_limit
 
 @pytest.mark.asyncio
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 async def test_create_sandbox_passes_gpu_device_requests(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -196,7 +196,7 @@ async def test_create_sandbox_passes_gpu_device_requests(mock_docker):
         patch.object(service, "_ensure_image_available"),
         patch.object(service, "_prepare_sandbox_runtime"),
         patch(
-            "opensandbox_server.services.docker.allocate_port_bindings",
+            "opensandbox_server.services.docker.docker_service.allocate_port_bindings",
             return_value={
                 "44772": ("0.0.0.0", 40001),
                 "8080": ("0.0.0.0", 40002),
@@ -215,7 +215,7 @@ async def test_create_sandbox_passes_gpu_device_requests(mock_docker):
     assert device_requests[0]["Capabilities"] == [["gpu"]]
 
 @pytest.mark.asyncio
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 async def test_create_sandbox_without_gpu_omits_device_requests(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -237,7 +237,7 @@ async def test_create_sandbox_without_gpu_omits_device_requests(mock_docker):
         patch.object(service, "_ensure_image_available"),
         patch.object(service, "_prepare_sandbox_runtime"),
         patch(
-            "opensandbox_server.services.docker.allocate_port_bindings",
+            "opensandbox_server.services.docker.docker_service.allocate_port_bindings",
             return_value={
                 "44772": ("0.0.0.0", 40001),
                 "8080": ("0.0.0.0", 40002),
@@ -268,7 +268,7 @@ async def test_create_sandbox_without_gpu_omits_device_requests(mock_docker):
     ],
 )
 @pytest.mark.asyncio
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 async def test_prepare_runtime_failure_triggers_cleanup(
     mock_docker, runtime_exc, expected_status, expect_wrapped_error
 ):
@@ -306,7 +306,7 @@ async def test_prepare_runtime_failure_triggers_cleanup(
         assert exc.value.detail["message"] == runtime_exc.detail["message"]
 
 @pytest.mark.asyncio
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 async def test_create_sandbox_rejects_invalid_metadata(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -331,7 +331,7 @@ async def test_create_sandbox_rejects_invalid_metadata(mock_docker):
     mock_client.containers.create.assert_not_called()
 
 @pytest.mark.asyncio
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 async def test_create_sandbox_rejects_timeout_above_configured_maximum(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -358,7 +358,7 @@ async def test_create_sandbox_rejects_timeout_above_configured_maximum(mock_dock
     assert "configured maximum of 3600s" in exc.value.detail["message"]
 
 @pytest.mark.asyncio
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 async def test_create_sandbox_rejects_unsupported_platform(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -382,7 +382,7 @@ async def test_create_sandbox_rejects_unsupported_platform(mock_docker):
     assert exc.value.detail["code"] == SandboxErrorCodes.INVALID_PARAMETER
     mock_client.containers.create.assert_not_called()
 
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 def test_ensure_image_available_repulls_when_cached_platform_mismatch(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -408,7 +408,7 @@ def test_ensure_image_available_repulls_when_cached_platform_mismatch(mock_docke
     assert call.args[3].os == "linux"
     assert call.args[3].arch == "arm64"
 
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 def test_ensure_image_available_repulls_when_platform_omitted_and_cached_arch_differs(
     mock_docker,
 ):
@@ -437,7 +437,7 @@ def test_ensure_image_available_repulls_when_platform_omitted_and_cached_arch_di
     assert call.args[3].os == "linux"
     assert call.args[3].arch == "amd64"
 
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 def test_ensure_image_available_does_not_repull_when_platform_omitted_and_cached_amd64(
     mock_docker,
 ):
@@ -461,7 +461,7 @@ def test_ensure_image_available_does_not_repull_when_platform_omitted_and_cached
 
     mock_pull.assert_not_called()
 
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 def test_pull_image_passes_platform_to_docker_api(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -481,7 +481,7 @@ def test_pull_image_passes_platform_to_docker_api(mock_docker):
         platform="linux/arm64",
     )
 
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 def test_pull_image_skips_platform_for_windows_profile(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -500,7 +500,7 @@ def test_pull_image_skips_platform_for_windows_profile(mock_docker):
         auth_config=None,
     )
 
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 def test_ensure_image_available_skips_windows_platform_mismatch_repull(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -522,7 +522,7 @@ def test_ensure_image_available_skips_windows_platform_mismatch_repull(mock_dock
 
     mock_pull.assert_not_called()
 
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 def test_fetch_execd_archive_caches_by_platform_key(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -552,7 +552,7 @@ def test_fetch_execd_archive_caches_by_platform_key(mock_docker):
     assert arm64_data == b"arm64"
     assert mock_client.containers.create.call_count == 2
 
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 def test_fetch_execd_archive_maps_platform_typeerror_to_invalid_parameter(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -569,7 +569,7 @@ def test_fetch_execd_archive_maps_platform_typeerror_to_invalid_parameter(mock_d
     assert "platform-aware container create" in exc_info.value.detail["message"]
 
 @pytest.mark.asyncio
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 async def test_create_sandbox_requires_entrypoint(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -595,7 +595,7 @@ async def test_create_sandbox_requires_entrypoint(mock_docker):
     mock_client.containers.create.assert_not_called()
 
 @pytest.mark.asyncio
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 async def test_network_policy_rejected_on_host_mode(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -623,7 +623,7 @@ async def test_network_policy_rejected_on_host_mode(mock_docker):
     assert exc.value.detail["code"] == SandboxErrorCodes.INVALID_PARAMETER
 
 @pytest.mark.asyncio
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 async def test_network_policy_requires_egress_image(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -651,7 +651,7 @@ async def test_network_policy_requires_egress_image(mock_docker):
     assert exc.value.detail["code"] == SandboxErrorCodes.INVALID_PARAMETER
 
 @pytest.mark.asyncio
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 async def test_egress_sidecar_injection_and_capabilities(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -683,9 +683,9 @@ async def test_egress_sidecar_injection_and_capabilities(mock_docker):
     )
 
     with (
-        patch("opensandbox_server.services.docker.generate_egress_token", return_value="egress-token"),
+        patch("opensandbox_server.services.docker.docker_service.generate_egress_token", return_value="egress-token"),
         patch(
-            "opensandbox_server.services.docker.allocate_port_bindings",
+            "opensandbox_server.services.docker.docker_service.allocate_port_bindings",
             return_value={
                 "44772": ("0.0.0.0", 44772),
                 "8080": ("0.0.0.0", 8080),
@@ -724,7 +724,7 @@ async def test_egress_sidecar_injection_and_capabilities(mock_docker):
 
 
 @pytest.mark.asyncio
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 async def test_create_sandbox_rejects_secure_access_on_docker_runtime(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -754,7 +754,7 @@ async def test_create_sandbox_rejects_secure_access_on_docker_runtime(mock_docke
 
 
 @pytest.mark.asyncio
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 async def test_network_policy_rejected_on_user_defined_network(mock_docker):
     """networkPolicy must be rejected when network_mode is a user-defined named network."""
     mock_client = MagicMock()
@@ -784,7 +784,7 @@ async def test_network_policy_rejected_on_user_defined_network(mock_docker):
     assert "my-custom-net" in exc.value.detail["message"]
 
 @pytest.mark.asyncio
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 async def test_create_sandbox_fails_when_user_defined_network_not_found(mock_docker):
     """create_sandbox raises 400 with a clear message when the named network does not exist."""
     from docker.errors import NotFound as DockerNotFound
@@ -816,7 +816,7 @@ async def test_create_sandbox_fails_when_user_defined_network_not_found(mock_doc
     assert "docker network create" in exc.value.detail["message"]
 
 @pytest.mark.asyncio
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 async def test_create_sandbox_user_defined_network_uses_correct_network_mode(mock_docker):
     """Containers created on a user-defined network use the network name as network_mode."""
 
@@ -848,7 +848,7 @@ async def test_create_sandbox_user_defined_network_uses_correct_network_mode(moc
         patch.object(service, "_ensure_image_available"),
         patch.object(service, "_prepare_sandbox_runtime"),
         patch(
-            "opensandbox_server.services.docker.allocate_port_bindings",
+            "opensandbox_server.services.docker.docker_service.allocate_port_bindings",
             return_value={
                 "44772": ("0.0.0.0", 40001),
                 "8080": ("0.0.0.0", 40002),
@@ -860,7 +860,7 @@ async def test_create_sandbox_user_defined_network_uses_correct_network_mode(moc
     call_kwargs = mock_client.api.create_container.call_args.kwargs
     assert call_kwargs["host_config"]["network_mode"] == "my-app-net"
 
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 def test_validate_network_skipped_for_builtin_modes(mock_docker):
     """_validate_network_exists does NOT call the Docker API for host or bridge modes."""
     mock_client = MagicMock()
@@ -875,7 +875,7 @@ def test_validate_network_skipped_for_builtin_modes(mock_docker):
         service._validate_network_exists()
         mock_client.networks.get.assert_not_called()
 
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 def test_egress_sidecar_cleanup_uses_api_remove_when_lookup_fails(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -916,7 +916,7 @@ def test_egress_sidecar_cleanup_uses_api_remove_when_lookup_fails(mock_docker):
     assert typed_detail["message"] == "Egress sidecar container failed to start."
     mock_client.api.remove_container.assert_called_once_with("sidecar-id", force=True)
 
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 def test_egress_sidecar_missing_id_preserves_specific_error(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -957,7 +957,7 @@ def test_egress_sidecar_missing_id_preserves_specific_error(mock_docker):
     mock_client.containers.get.assert_not_called()
     mock_client.api.remove_container.assert_not_called()
 
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 def test_egress_sidecar_cleanup_wraps_unexpected_lookup_error(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -999,7 +999,7 @@ def test_egress_sidecar_cleanup_wraps_unexpected_lookup_error(mock_docker):
     assert typed_detail["message"] == "Egress sidecar container failed to start."
     mock_client.api.remove_container.assert_called_once_with("sidecar-id", force=True)
 
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 def test_egress_sidecar_host_config_sysctls_only_when_egress_disable_ipv6(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -1056,7 +1056,7 @@ def test_egress_sidecar_host_config_sysctls_only_when_egress_disable_ipv6(mock_d
     assert hc2["sysctls"]["net.ipv6.conf.all.disable_ipv6"] == 1
 
 
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 def test_egress_sidecar_normalizes_windows_port_bindings(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -1222,7 +1222,7 @@ def test_build_labels_store_platform_constraints():
     assert labels[SANDBOX_PLATFORM_ARCH_LABEL] == "arm64"
 
 @pytest.mark.asyncio
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 async def test_create_sandbox_with_manual_cleanup_completes_full_create_path(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -1250,7 +1250,7 @@ async def test_create_sandbox_with_manual_cleanup_completes_full_create_path(moc
     mock_schedule.assert_not_called()
 
 @pytest.mark.asyncio
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 async def test_create_sandbox_passes_platform_to_container_create(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -1273,7 +1273,7 @@ async def test_create_sandbox_passes_platform_to_container_create(mock_docker):
     assert called_args[-1].arch == "arm64"
 
 @pytest.mark.asyncio
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 async def test_create_sandbox_response_keeps_platform_null_when_unconstrained(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -1297,7 +1297,7 @@ async def test_create_sandbox_response_keeps_platform_null_when_unconstrained(mo
 
     assert response.platform is None
 
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 def test_create_and_start_container_uses_unconstrained_platform_for_execd(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -1328,7 +1328,7 @@ def test_create_and_start_container_uses_unconstrained_platform_for_execd(mock_d
     assert passed_platform.os == "linux"
     assert passed_platform.arch == "arm64"
 
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 def test_create_and_start_container_maps_platform_typeerror_to_invalid_parameter(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -1354,7 +1354,7 @@ def test_create_and_start_container_maps_platform_typeerror_to_invalid_parameter
     assert "platform-aware container create" in exc_info.value.detail["message"]
 
 
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 def test_create_and_start_container_windows_profile_keeps_image_entrypoint(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -1369,9 +1369,9 @@ def test_create_and_start_container_windows_profile_keeps_image_entrypoint(mock_
 
     service = DockerSandboxService(config=_app_config())
     with (
-        patch("opensandbox_server.services.docker.fetch_execd_install_bat", return_value=b"script"),
-        patch("opensandbox_server.services.docker.fetch_execd_windows_binary", return_value=b"exe"),
-        patch("opensandbox_server.services.docker.install_windows_oem_scripts") as mock_install,
+        patch("opensandbox_server.services.docker.container_ops.fetch_execd_install_bat", return_value=b"script"),
+        patch("opensandbox_server.services.docker.container_ops.fetch_execd_windows_binary", return_value=b"exe"),
+        patch("opensandbox_server.services.docker.container_ops.install_windows_oem_scripts") as mock_install,
     ):
         service._create_and_start_container(
             sandbox_id="sandbox-win-1",
@@ -1391,7 +1391,7 @@ def test_create_and_start_container_windows_profile_keeps_image_entrypoint(mock_
     mock_install.assert_called_once()
 
 
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 def test_create_and_start_container_windows_profile_skips_linux_runtime_injection(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -1406,9 +1406,9 @@ def test_create_and_start_container_windows_profile_skips_linux_runtime_injectio
     service = DockerSandboxService(config=_app_config())
     with (
         patch.object(service, "_prepare_sandbox_runtime") as mock_prepare,
-        patch("opensandbox_server.services.docker.fetch_execd_install_bat", return_value=b"script"),
-        patch("opensandbox_server.services.docker.fetch_execd_windows_binary", return_value=b"exe"),
-        patch("opensandbox_server.services.docker.install_windows_oem_scripts") as mock_install,
+        patch("opensandbox_server.services.docker.container_ops.fetch_execd_install_bat", return_value=b"script"),
+        patch("opensandbox_server.services.docker.container_ops.fetch_execd_windows_binary", return_value=b"exe"),
+        patch("opensandbox_server.services.docker.container_ops.install_windows_oem_scripts") as mock_install,
     ):
         service._create_and_start_container(
             sandbox_id="sandbox-win-2",
@@ -1426,7 +1426,7 @@ def test_create_and_start_container_windows_profile_skips_linux_runtime_injectio
 
 
 @pytest.mark.asyncio
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 async def test_create_sandbox_windows_profile_injects_runtime_defaults(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -1447,7 +1447,7 @@ async def test_create_sandbox_windows_profile_injects_runtime_defaults(mock_dock
 
     with (
         patch(
-            "opensandbox_server.services.docker.validate_windows_runtime_prerequisites",
+            "opensandbox_server.services.docker.docker_service.validate_windows_runtime_prerequisites",
             return_value=[],
         ),
         patch.object(
@@ -1474,7 +1474,7 @@ async def test_create_sandbox_windows_profile_injects_runtime_defaults(mock_dock
 
 
 @pytest.mark.asyncio
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 async def test_create_sandbox_windows_profile_does_not_require_download_url_override(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -1494,7 +1494,7 @@ async def test_create_sandbox_windows_profile_does_not_require_download_url_over
 
     with (
         patch(
-            "opensandbox_server.services.docker.validate_windows_runtime_prerequisites",
+            "opensandbox_server.services.docker.docker_service.validate_windows_runtime_prerequisites",
             return_value=[],
         ),
         patch.object(
@@ -1509,7 +1509,7 @@ async def test_create_sandbox_windows_profile_does_not_require_download_url_over
 
 
 @pytest.mark.asyncio
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 async def test_create_sandbox_windows_profile_rejects_missing_runtime_devices(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -1527,7 +1527,7 @@ async def test_create_sandbox_windows_profile_rejects_missing_runtime_devices(mo
     )
     with (
         patch(
-            "opensandbox_server.services.docker.validate_windows_runtime_prerequisites",
+            "opensandbox_server.services.docker.docker_service.validate_windows_runtime_prerequisites",
             side_effect=HTTPException(
                 status_code=status.HTTP_400_BAD_REQUEST,
                 detail={
@@ -1548,7 +1548,7 @@ async def test_create_sandbox_windows_profile_rejects_missing_runtime_devices(mo
 
 
 @pytest.mark.asyncio
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 async def test_create_sandbox_windows_profile_rejects_below_minimum_resource_limits(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -1566,7 +1566,7 @@ async def test_create_sandbox_windows_profile_rejects_below_minimum_resource_lim
     )
     with (
         patch(
-            "opensandbox_server.services.docker.validate_windows_runtime_prerequisites",
+            "opensandbox_server.services.docker.docker_service.validate_windows_runtime_prerequisites",
             return_value=None,
         ),
         patch.object(service, "_create_and_start_container") as mock_create,
@@ -1581,7 +1581,7 @@ async def test_create_sandbox_windows_profile_rejects_below_minimum_resource_lim
 
 
 @pytest.mark.asyncio
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 async def test_create_sandbox_windows_profile_accepts_dockur_demo_like_request(mock_docker):
     """
     Use a dockur/windows-style request payload (VERSION env) and verify
@@ -1613,7 +1613,7 @@ async def test_create_sandbox_windows_profile_accepts_dockur_demo_like_request(m
 
     with (
         patch(
-            "opensandbox_server.services.docker.validate_windows_runtime_prerequisites",
+            "opensandbox_server.services.docker.docker_service.validate_windows_runtime_prerequisites",
             return_value=None,
         ),
         patch.object(
@@ -1639,7 +1639,7 @@ async def test_create_sandbox_windows_profile_accepts_dockur_demo_like_request(m
 
 
 @pytest.mark.asyncio
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 async def test_create_sandbox_windows_profile_with_network_policy_maps_windows_ports(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -1671,12 +1671,12 @@ async def test_create_sandbox_windows_profile_with_network_policy_maps_windows_p
 
     with (
         patch(
-            "opensandbox_server.services.docker.validate_windows_runtime_prerequisites",
+            "opensandbox_server.services.docker.docker_service.validate_windows_runtime_prerequisites",
             return_value=None,
         ),
-        patch("opensandbox_server.services.docker.generate_egress_token", return_value="egress-token"),
+        patch("opensandbox_server.services.docker.docker_service.generate_egress_token", return_value="egress-token"),
         patch(
-            "opensandbox_server.services.docker.allocate_port_bindings",
+            "opensandbox_server.services.docker.docker_service.allocate_port_bindings",
             return_value={
                 "44772": ("0.0.0.0", 51664),
                 "8080": ("0.0.0.0", 48891),
@@ -1731,7 +1731,7 @@ def test_restore_existing_sandboxes_ignores_manual_cleanup_without_warning():
 
     with (
         patch.object(service.docker_client.containers, "list", return_value=[manual_container]),
-        patch("opensandbox_server.services.docker.logger.warning") as mock_warning,
+        patch("opensandbox_server.services.docker.docker_service.logger.warning") as mock_warning,
         patch.object(service, "_schedule_expiration") as mock_schedule,
     ):
         service._restore_existing_sandboxes()
@@ -1739,7 +1739,7 @@ def test_restore_existing_sandboxes_ignores_manual_cleanup_without_warning():
     mock_schedule.assert_not_called()
     mock_warning.assert_not_called()
 
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 def test_pending_snapshot_restore_reports_snapshot_id_without_image(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -1764,7 +1764,7 @@ def test_pending_snapshot_restore_reports_snapshot_id_without_image(mock_docker)
     assert sandbox.snapshot_id == "snap-001"
     assert sandbox.image is None
 
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 def test_container_snapshot_restore_reports_snapshot_id_without_image(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -1795,7 +1795,7 @@ def test_container_snapshot_restore_reports_snapshot_id_without_image(mock_docke
     assert sandbox.snapshot_id == "snap-001"
     assert sandbox.image is None
 
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 def test_delete_sandbox_removes_windows_oem_volume(mock_docker):
     mock_container = MagicMock()
     mock_container.attrs = {
@@ -1818,7 +1818,7 @@ def test_delete_sandbox_removes_windows_oem_volume(mock_docker):
     mock_client.api.remove_volume.assert_called_once_with("opensandbox-win-oem-sandbox-win-1")
 
 
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 def test_delete_sandbox_skips_oem_volume_cleanup_for_linux(mock_docker):
     mock_container = MagicMock()
     mock_container.attrs = {
@@ -1861,7 +1861,7 @@ def test_renew_expiration_rejects_manual_cleanup_sandbox():
     assert exc_info.value.detail["message"] == "Sandbox manual-id does not have automatic expiration enabled."
 
 @pytest.mark.asyncio
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 async def test_create_sandbox_async_returns_provisioning(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -1899,7 +1899,7 @@ async def test_create_sandbox_async_returns_provisioning(mock_docker):
     mock_sync.assert_called_once()
 
 @pytest.mark.asyncio
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 async def test_get_sandbox_returns_pending_state(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -1935,7 +1935,7 @@ async def test_get_sandbox_returns_pending_state(mock_docker):
     assert response.status.state == "Running"
     assert response.entrypoint == ["python", "app.py"]
 
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 def test_list_sandboxes_deduplicates_container_and_pending(mock_docker):
     # Build a realistic container mock to avoid parse_timestamp errors.
     container = MagicMock()
@@ -1982,7 +1982,7 @@ def test_list_sandboxes_deduplicates_container_and_pending(mock_docker):
     assert response.items[0].status.state == "Running"
     assert response.items[0].metadata == {"team": "c"}
 
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 def test_get_sandbox_prefers_container_over_pending(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -2026,7 +2026,7 @@ def test_get_sandbox_prefers_container_over_pending(mock_docker):
     assert sandbox.status.state == "Running"
     assert sandbox.entrypoint == ["/bin/sh"]
 
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 def test_async_worker_cleans_up_leftover_container_on_failure(mock_docker):
     mock_client = MagicMock()
     mock_client.containers.list.return_value = []
@@ -2068,7 +2068,7 @@ def test_async_worker_cleans_up_leftover_container_on_failure(mock_docker):
     service._cleanup_failed_containers.assert_called_once_with(sandbox_id)
     assert service._pending_sandboxes[sandbox_id].status.state == "Failed"
 
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 class TestBuildVolumeBinds:
 
     def test_none_volumes_returns_empty(self, mock_docker):
@@ -2257,7 +2257,7 @@ class TestBuildVolumeBinds:
         binds = service._build_volume_binds([volume])
         assert binds == ["/mnt/ossfs/bucket-test-3/task-001:/mnt/data:rw"]
 
-@patch("opensandbox_server.services.docker.docker")
+@patch("opensandbox_server.services.docker.docker_service.docker")
 class TestDockerVolumeValidation:
 
     @pytest.mark.asyncio
@@ -2952,7 +2952,7 @@ class TestDockerVolumeValidation:
 
         # Simulate: realpath resolves a symlink that escapes the mountpoint.
         # datasets -> / inside the volume, so realpath(…/_data/datasets) = /
-        with patch("opensandbox_server.services.docker.os.path.realpath") as mock_realpath:
+        with patch("opensandbox_server.services.docker.docker_service.os.path.realpath") as mock_realpath:
             mock_realpath.side_effect = lambda p, **kwargs: ("/" if p.endswith("datasets") else p)
             with pytest.raises(HTTPException) as exc_info:
                 await service.create_sandbox(request)
@@ -3037,7 +3037,7 @@ class TestDockerVolumeValidation:
             ],
         )
 
-        with patch("opensandbox_server.services.docker.os.makedirs", side_effect=PermissionError("denied")):
+        with patch("opensandbox_server.services.docker.docker_service.os.makedirs", side_effect=PermissionError("denied")):
             with pytest.raises(HTTPException) as exc_info:
                 await service.create_sandbox(request)
 
@@ -3430,7 +3430,7 @@ class TestDockerVolumeValidation:
 def test_docker_get_endpoint_rejects_expires():
     from unittest.mock import patch
 
-    with patch("opensandbox_server.services.docker.docker"):
+    with patch("opensandbox_server.services.docker.docker_service.docker"):
         cfg = _app_config()
         cfg.docker.network_mode = "bridge"
         service = DockerSandboxService(config=cfg)
